% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{access}
\alias{access}
\title{Return accessible projects and receivers}
\usage{
access()
}
\description{
Return the projects and receivers which are accessible by the given
credentials
}
\examples{
\dontrun{
access()
}

}
