% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Add/update batch activity}
\usage{
activity(src, resume = FALSE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{resume}{Logical. Resume a download? Otherwise the activity table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the 'activity' table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\details{
This function is automatically run by the \code{\link[=tagme]{tagme()}} function with
\code{resume = TRUE}.

If an 'activity' table doesn't exist, it will be created prior to
downloading. If there is an existing 'activity' table, this will update the
records.
}
\examples{
\dontrun{
  # First download tag data
  my_tags <- tagme(176, new = TRUE)
  
  # Now download batch activity 
  my_tags <- activity(my_tags)
  
  # If interrupted and you want to resume
  my_tags <- activity(my_tags, resume = TRUE)
}

}
