% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activityAll.R
\name{activityAll}
\alias{activityAll}
\title{Add/update all batch activity}
\usage{
activityAll(src, resume = FALSE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}

\item{resume}{Logical. Resume a download? Otherwise the table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the \code{activityAll} table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()
  
# Get all activity
\dontrun{sql_motus <- activityAll(sql_motus)}

# Access 'activityAll' table
library(dplyr)
a <- tbl(sql_motus, "activityAll")
  
# If interrupted and you want to resume
\dontrun{sql_motus <- activityAll(sql_motus, resume = TRUE)}

}
