% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVersion.R
\name{checkVersion}
\alias{checkVersion}
\title{Check database version}
\usage{
checkVersion(src)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}}
}
\description{
Verifies the version of the package against the \code{admInfo} table of a \code{.motus}
file. Those should match if the \code{updateMotusDb()} function has been properly
applied by the \code{tagme()} function.
}
