% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consec_site.fun.R
\name{consec.fun}
\alias{consec.fun}
\title{Create dataframe for siteTrans function}
\usage{
consec.fun(df)
}
\arguments{
\item{data}{dataframe of Motus detection data containing at a minimum fullID, ts, lat, lon}
}
\description{
Create dataframe for siteTrans function
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
