% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRunsFilter.R
\name{createRunsFilter}
\alias{createRunsFilter}
\title{Create a new filter records that can be applied to runs}
\usage{
createRunsFilter(src, filterName, motusProjID = NA, descr = NA, update = FALSE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}}

\item{filterName}{Character. Unique name given to the filter}

\item{motusProjID}{Character. Optional project ID attached to the filter in
order to share with other users of the same project.}

\item{descr}{Character. Optional filter description detailing what the filter
is meant to do}

\item{update}{Logical. Whether the filter record gets updated when a filter
with the same name already exists.}
}
\value{
an integer \code{filterID}
}
\description{
Create a new filter records that can be applied to runs
}
