% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteRunsFilter.R
\name{deleteRunsFilter}
\alias{deleteRunsFilter}
\title{Delete a filter}
\usage{
deleteRunsFilter(src, filterName, motusProjID = NA, clearOnly = FALSE)
}
\arguments{
\item{src}{SQLite connection (result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")})}

\item{filterName}{Character. Unique name given to the filter}

\item{motusProjID}{Character. Optional project ID attached to the filter in
order to share with other users of the same project.}

\item{clearOnly}{Logical. When true, only remove the probability records
associated with the filter, but retain the filter itself}
}
\value{
the integer \code{filterID} of the filter deleted
}
\description{
Deletes a filter by name or project ID.
}
