% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecateBatches.R
\name{deprecateBatches}
\alias{deprecateBatches}
\title{Fetch and remove deprecated batches}
\usage{
deprecateBatches(src, fetchOnly = FALSE, ask = TRUE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{fetchOnly}{Logical. Only \emph{fetch} batches that are deprecated. Don't
remove deprecated batches from other tables.}

\item{ask}{Logical. Ask for confirmation when removing deprecated batches}
}
\description{
Deprecated batches are removed from the online database but not from
local data files. This function fetches a list of deprecated batches
(stored in the 'deprecated' table), and, optionally, removes these batches
from all tables that reference \code{batchID}s
}
\examples{

# Download and access data from project 176 in sql format
\dontrun{
# usename and password are both "motus.sample"
sql.motus <- tagme(176, new = TRUE, update = TRUE)

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
  
# Access 'deprecated' table using tbl() from dplyr
library(dplyr)
tbl(sql.motus, "deprecated")

# See that there are deprecated batches in the data
filter(tbl(sql.motus, "alltags"), batchID == 6000)

# Fetch deprecated batches
deprecateBatches(sql.motus, fetchOnly = TRUE)

# Remove deprecated batches (will ask for confirmation unless ask = FALSE)
deprecateBatches(sql.motus, ask = FALSE)

# See that there are NO more deprecated batches in the data
filter(tbl(sql.motus, "alltags"), batchID == 6000)
}
}
