% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Initialize constants and writable storage for the motus package}
\usage{
.onLoad(...)
}
\description{
This function initializes the motus_vars environment that holds
constants (whose bindings are locked) and session variables (not
locked) used by this package.
}
\details{
Constants are:

\describe{
\item{API_API_INFO}{URL to call for getting information about the (data) API}
\item{API_REGISTER_TAG}{URL to call for registering a tag}
\item{API_DEPLOY_TAG}{URL to call for deploying a tag}
\item{API_SEARCH_TAGS}{URL to call for listing registered tags}
\item{API_RECEIVERS_FOR_PROJECT}{URL to call for getting a list of receiver deployments for a project}
\item{API_DEVICE_ID_FOR_RECEIVER}{URL to call for getting the device ID for a receiver}
\item{API_BATCHES_FOR_TAG_PROJECT}{URL to call for getting batches for a tag project}
\item{API_BATCHES_FOR_RECEIVER}{URL to call for getting batches for a receiver}
\item{API_BATCHES_FOR_ALL}{URL to call for getting batches for any receiver}
\item{API_GPS_FOR_TAG_PROJECT}{URL to call for getting GPS fixes for a tag project}
\item{API_GPS_FOR_RECEIVER}{URL to call for getting GPS fixes for a receiver project}
\item{API_PULSE_COUNTS_FOR_RECEIVER}{URL to call for getting antenna pulse counts for a receiver}
\item{API_METADATA_FOR_TAGS}{URL to call for getting metadata for tags}
\item{API_METADATA_FOR_RECEIVERS}{URL to call for getting metadata for receivers}
\item{API_RECV_METADATA_FOR_PROJECTS}{URL to call for getting receiver metadata by projects}
\item{API_TAG_METADATA_FOR_PROJECTS}{URL to call for getting tag metadata by projects}
\item{API_TAGS_FOR_AMBIGIUITIES}{URL to call for getting motus tagIDs represented by an ambiguity ID}
\item{API_PROJECT_AMBIGUITIES_FOR_TAG_PROJECT}{URL to call for getting list of ambiguous project IDs for a project}
\item{FLOAT_FIELDS}{list of API fieldnames requiring floating point values}
\item{FLOAT_REGEX}{regex to recognize fields requiring fixups in API queries}
}

and variables are:

\describe{
\item{userLogin}{login name for user at motus.org}
\item{userPassword}{password for user at motus.org}
\item{projects}{project IDs for user at motus.org}
\item{dataServerURL}{URL to data server}
\item{dbDir}{path to folder with project and tag databases}
}
}
\seealso{
\code{\link{motus_vars}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\keyword{internal}
