% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAccess}
\alias{getAccess}
\title{Return accessible projects and receivers}
\usage{
getAccess()
}
\description{
Return the projects and receivers which are accessible by the given
credentials
}
\examples{
\dontrun{
getAccess()
}

}
