% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBluPayload.R
\name{getBluPayload}
\alias{getBluPayload}
\title{Extract variables from BLUtag Payloads}
\usage{
getBluPayload(df_src)
}
\arguments{
\item{df_src}{Data frame, SQLite connection, or SQLite table. An SQLite
connection would be the result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}; an SQLite table would be
the result of \code{dplyr::tbl(tags, "alltags")}; a data frame could be the
result of \code{dplyr::tbl(tags, "alltags") \%>\% dplyr::collect()}.}
}
\value{
Data frame/Tibble of the BLUtag hits with extra variable columns.
}
\description{
Extracts the data from the hexadecimal BLUtag payload into new columns.
}
\examples{
# With example, dummy data
t <- data.frame(hitID = 99999999, batchID = 9999999, sync = 49000, product = 1, revision = 0)
t <- cbind(t, payload = c("6406060F", "6706150F", "6106270F", NA))
t 

# Extract payload
getBluPayload(t)

\dontrun{
  t <- tagme(XXX, update = TRUE, new = TRUE) # Hits with BLUTags
  getBluPayload(t) # Extract payload from the database
  getBluPayload(dplyr::tbl(t, "hitsBlu")) # Or supply the hitsBlu table directly
}
}
