% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDBFilename.R
\name{getDBFilename}
\alias{getDBFilename}
\title{get the full filename for a tag project or receiver database}
\usage{
getDBFilename(projRecv, dir = getwd())
}
\arguments{
\item{projRecv}{integer scalar project code from motus.org, *or*
character scalar receiver serial number.}

\item{dir}{character scalar: path to folder where databases are stored
default: \code{getwd()}}
}
\value{
character scalar giving the full path to the database file for
the specified tag project or receiver.  Both DB types have file extensions
".motus", and the base filename looks like either "project-X.motus" or
"SG-1234BBBK5678.motus" or "Lotek-12345.motus"
}
\description{
get the full filename for a tag project or receiver database
}
\examples{

## getDBFilename(123) # get the path to the database for tag project 123

## getDBFilename("SG-1234BBBK5678") # get the path to the database
##     for receiver SG-1234BBBK5678

## getDBFilename(5, "/home/me/mydbs") # get the path to the
##       database for project 5, which is in folder \\code{/home/me/mydbs}

}
\seealso{
\code{\link{tagme}}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
