% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMotusDBSrc.R
\name{getMotusDBSrc}
\alias{getMotusDBSrc}
\title{Get the src_sqlite for a receiver or tag database}
\usage{
getMotusDBSrc(
  recv = NULL,
  proj = NULL,
  create = FALSE,
  dbDir = motus_vars$dbDir
)
}
\arguments{
\item{recv}{receiver serial number}

\item{proj}{integer motus project number
exactly  one of \code{proj} or \code{recv} must be specified.}

\item{create}{Is this a new database?  Default: FALSE. Same semantics as for
\code{src_sqlite()}'s parameter of the same name:  the DB must already exist
unless you specify \code{create = TRUE}}

\item{dbDir}{path to folder with existing receiver databases Default:
\code{motus_vars$dbDir}, which is set to the current folder by
\code{getwd()} when this library is loaded.}
}
\value{
a src_sqlite for the receiver; if the receiver is new, this database
will be empty, but have the correct schema.
}
\description{
Receiver database files have names like "SG-1234BBBK06EA.motus" or
"Lotek-12345.motus", and project database files have names like
"project-52.motus".
}
