% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunsFilterID.R
\name{getRunsFilterID}
\alias{getRunsFilterID}
\title{Returns the filterID matching a filter name (and optionally a project ID)
If no project ID is specified and there is only 1 filter of the correct name,
return the filterID regardless of the project it is attached to.
If a project ID is specified or there are more than 1 filters of the name,
return the specific one matching the project ID (including NA).}
\usage{
getRunsFilterID(src, filterName, motusProjID = NA)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{filterName}{unique name given to the filter}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{descr}{optional filter description detailing what the filter is meant to do}
}
\value{
an integer filterID
}
\description{
Returns the filterID matching a filter name (and optionally a project ID)
If no project ID is specified and there is only 1 filter of the correct name,
return the filterID regardless of the project it is attached to.
If a project ID is specified or there are more than 1 filters of the name,
return the specific one matching the project ID (including NA).
}
\author{
Denis Lepage, Bird Studies Canada
}
