% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsAll.R
\name{gpsAll}
\alias{gpsAll}
\title{Add/update all GPS points}
\usage{
gpsAll(src, resume = TRUE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}}

\item{resume}{Logical. Resume a download? Otherwise the table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the \code{gpsAll} table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE, update = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()
  
# Get all GPS points
\dontrun{sql_motus <- gpsAll(sql_motus)}

# Access 'gpsAll' table
library(dplyr)
g <- tbl(sql_motus, "gpsAll")
  
# gpsAll resumes a previous download by default
# If you want to delete this original data and do a fresh download, 
# use resume = FALSE
\dontrun{sql_motus <- gpsAll(sql_motus, resume = FALSE)}

}
