% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRunsFilters.R
\name{listRunsFilters}
\alias{listRunsFilters}
\title{Returns a dataframe of the filters stored in the local database.}
\usage{
listRunsFilters(src)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}
}
\value{
a dataframe
}
\description{
Returns a dataframe of the filters stored in the local database.
}
