% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAllambigsView.R
\name{makeAllambigsView}
\alias{makeAllambigsView}
\title{create a virtual table that transposes the tagAmbig table}
\usage{
makeAllambigsView(db, name = "allambigs")
}
\arguments{
\item{db}{dplyr src_sqlite to detections database}

\item{name}{character scalar; name for the virtual table.
Default: 'allambigs'.}
}
\value{
a dplyr::tbl which refers to the newly-created virtual table.
By default, the columns in the virtual table are:
\itemize{
   \item{ambigID} unique ID linking ambiguous tag detections
   \item{motusTagID} unique motus ID for this physical tag
}
}
\description{
Creates a virtual table (really a 'view') in a motus database that
converts the list of tags associated with ambiguous detections from columns to rows
this is mainly used by the alltags view to expand the detections records showing each tag
in addition to the ambigID
}
\note{
The new virtual table replaces any previous virtual table by the same
name in \code{db}.  The virtual table is an SQL VIEW, which will persist in \code{db}
across R sessions.
}
\author{
Denis Lepage
}
