% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSimpletagsView.R
\name{makeSimpletagsView}
\alias{makeSimpletagsView}
\title{create a view of tag detections with basic metadata}
\usage{
makeSimpletagsView(db, name = "simpletags")
}
\arguments{
\item{db}{dplyr src_sqlite to detections database}

\item{name}{character scalar; name for the virtual table.
Default: 'simpletags'.}
}
\value{
a dplyr::tbl which refers to the newly-created virtual table.
The columns in the virtual table are:
\itemize{
   \item hitID numeric; unique motus identifier for this detection
   \item motusTagID integer; unique motus identifier for the tag detected
   \item mfgID numeric; manufacturer tag ID
   \item fullID character; full ID of tag, in form "PROJ#MFGID:BI@FREQ"
   \item ts numeric; timestamp of detection; seconds since 1 Jan 1970 00:00:00 GMT
   \item sig numeric; relative signal strength (dB max)
   \item sigSD numeric; standard deviation of signal strength (dB)
   \item noise numeric; relative noise level (dB max)
   \item freq numeric; offset frequency (kHz)
   \item freqSD numeric; standard deviation of offset frequency
   \item slop numeric; total absolute deviation from true tag pulse spacing (milliseconds)
   \item burstSlop numeric; deviation from registered tag burst interval (seconds)
   \item runID numeric; unique motus ID for run of detections including this one
   \item ant character; antenna identifier (usually a small integer, for USB port)
   \item runLen integer; number of detections in this run
   \item motusDeviceID integer; unique motus ID for receiver that detected tag
   \item monoBN integer; boot number
   \item tagFreq numeric; nominal tag frequency (MHz)
   \item tagProjID integer; unique motus identifier for project that deployed tag
   \item tagProj character; short name of motus project that deployed tag
   \item speciesID integer; unique species identifier (from motus.org)
   \item recv character; receiver serial number
   \item site character; receiver site name
   \item lat numeric; latitude in decimal degrees North (so negative is South)
   \item long numeric; longitude in decimal degrees East (so negative is West)
   \item alt numeric; altitude in metres above sea level
   \item recvProjID; integer unique ID of motus project that deployed receiver
   \item recvProj; character; short name of motus project that deployed receiver
}
}
\description{
Creates a virtual table (really a 'view') in a motus database that
links each tag detection to basic metadata for the tag and
receiver.  This is a much "narrower" view than the 'alltags' view
created by \code{makeAlltagsView()}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
