% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Update all metadata}
\usage{
metadata(src, projectIDs = NULL, replace = TRUE, delete = FALSE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}

\item{projectIDs}{optional integer vector of Motus projects IDs for which
metadata should be obtained; default: NULL, meaning obtain metadata for all
tags and receivers that your permissions allow.}

\item{replace}{logical scalar; if TRUE (default), existing data replace the
existing metadata with the newly acquired ones.}

\item{delete}{logical scalar; Default = FALSE. if TRUE, the entire metadata
tables are cleared (for all projects) before re-importing the metadata.}
}
\description{
Updates the entire metadata for receivers and tags from Motus server.
Contrary to \code{tagme()}, this function retrieves the entire set of metadata for
tags and receivers, and not only those pertinent to the detections in your
local file.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()
                   
# Add extended metadata to your file
\dontrun{metadata(sql_motus)}
  
# Access different metadata tables
library(dplyr)
tbl(sql_motus, "species")
tbl(sql_motus, "projs")
tbl(sql_motus, "tagDeps")
# Etc.
  
}
\seealso{
\code{tagme()} provides an option to update only the metadata
relevant to a specific project or receiver file.
}
