% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motus-pkg.R
\docType{package}
\name{motus}
\alias{motus-package}
\alias{motus}
\title{Fetch and use data from the Motus Wildlife Tracking System}
\description{
\code{motus} is an R package for retrieving telemetry data from the Motus
Wildlife Tracking System \url{https://motus.org}.
}
\details{
For a detailed walk-though and instructions check out the
\href{https://motuswts.github.io/motus/}{walk-throughs and articles}!

Commonly used functions:
\enumerate{
\item Download telemetry data
\itemize{
\item \code{\link[=tagme]{tagme()}}
\item \code{\link[=tellme]{tellme()}}
\item \code{\link[=metadata]{metadata()}}
\item \code{\link[=checkVersion]{checkVersion()}}
}
\item Create data filters
\itemize{
\item \code{\link[=listRunsFilters]{listRunsFilters()}}
\item \code{\link[=getRunsFilters]{getRunsFilters()}}
\item \code{\link[=createRunsFilter]{createRunsFilter()}}
\item \code{\link[=writeRunsFilter]{writeRunsFilter()}}
\item \code{\link[=deleteRunsFilter]{deleteRunsFilter()}}
}
\item Summarize data
\itemize{
\item \code{\link[=tagSum]{tagSum()}}
\item \code{\link[=tagSumSite]{tagSumSite()}}
\item \code{\link[=simSiteDet]{simSiteDet()}}
\item \code{\link[=siteSum]{siteSum()}}
\item \code{\link[=siteSumDaily]{siteSumDaily()}}
\item \code{\link[=siteTrans]{siteTrans()}}
}
\item Plot data
\itemize{
\item \code{\link[=plotAllTagsCoord]{plotAllTagsCoord()}}
\item \code{\link[=plotAllTagsSite]{plotAllTagsSite()}}
\item \code{\link[=plotDailySiteSum]{plotDailySiteSum()}}
\item \code{\link[=plotRouteMap]{plotRouteMap()}}
\item \code{\link[=plotSite]{plotSite()}}
\item \code{\link[=plotSiteSig]{plotSiteSig()}}
\item \code{\link[=plotTagSig]{plotTagSig()}}
}
\item Sunrises and sets
\itemize{
\item \code{\link[=sunRiseSet]{sunRiseSet()}}
\item \code{\link[=timeToSunriset]{timeToSunriset()}}
}
}
}
\references{
Motus Wildlife Tracking System \url{https://motus.org}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://motuswts.github.io/motus/}
  \item Report bugs at \url{https://github.com/MotusWTS/motus/issues}
}

}
\author{
\strong{Maintainer}: Birds Canada \email{motus@birdscanada.org}

Authors:
\itemize{
  \item John Brzustowski \email{jbrzusto@fastmail.fm}
  \item Denis Lepage \email{dlepage@birdscanada.org}
}

Other contributors:
\itemize{
  \item Steffi LaZerte \email{sel@steffilazerte.ca} [contributor]
  \item Joey Bernard [contributor]
  \item Lucas Berrigan [contributor]
  \item Tara Crewe [contributor]
  \item Zoe Crysler \email{zcrysler@gmail.com} [contributor]
  \item Jeremy Hussell [contributor]
  \item Catherine Jardine [contributor]
  \item Amie MacDonald [contributor]
  \item Stuart Mackenzie [contributor]
  \item Paul Morrill [contributor]
  \item Josh Sayers [contributor]
  \item Philip Taylor [contributor]
}

}
\keyword{internal}
