% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusLogout.R
\name{motusLogout}
\alias{motusLogout}
\title{Forget login credentials for motus.}
\usage{
motusLogout()
}
\value{
TRUE.
}
\description{
Any requests to the motus data server after calling
this function will require re-entering a username and
password.
}
\details{
This function just resets these items to NULL:
\itemize{
\item motus_vars$authToken
\item motus_vars$userLogin
\item motus_vars$userPassword
}
Due to their active bindings, subsequent calls to
any functions that need them will prompt for a login.
}
