% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusQuery.R
\name{motusQuery}
\alias{motusQuery}
\title{Send a query to the motus API}
\usage{
motusQuery(API, params = NULL, requestType = "post", show = FALSE,
  JSON = FALSE)
}
\arguments{
\item{API}{one of the Motus$API_... constants}

\item{params}{named list of API-specific parameters}

\item{requestType}{"post" or "get"; default: "post".}

\item{show}{if TRUE, print the request to the console before submitting to motus}

\item{JSON}{if TRUE, return results as JSON-format string; otherwise, as R list}
}
\value{
the result of sending the request to the motus API.  The
    result is a JSON-format character scalar if \code{json} is
    \code{TRUE}; otherwise it is an R list with named components,
    extracted from the JSON return value.
}
\description{
Send a query to the motus API
}
\note{
If you have not already entered your motus login credentials
    in this R session, e.g. by calling \code{\link{motusLogin()}},
    this function will prompt you for them.
}
\author{
John Brzustowski
    \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
