% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusSearchTags.R
\name{motusSearchTags}
\alias{motusSearchTags}
\title{return a list of motus tags}
\usage{
motusSearchTags(projectID = NULL, tsStart = NULL, tsEnd = NULL,
  searchMode = c("startsBetween", "overlaps"), defaultLifespan = 90,
  lifespanBuffer = 1.5, regStart = NULL, regEnd = NULL, mfgID = NULL,
  ...)
}
\arguments{
\item{projectID:}{integer scalar; motus internal project ID}

\item{tsStart:}{numeric scalar; start of active period}

\item{tsEnd:}{numeric scalar; end of active period}

\item{searchMode:}{character scalar; type of search
desired. "overlaps" looks for tags active during at least a
portion of the time span \code{c(tsStart, tsEnd)}, while
"startsBetween" looks for tags with deployment start times in
the same range.}

\item{defaultLifeSpan:}{integer scalar; default lifespan of tags,
in days; used when motus does not know the lifespan for a tag.}

\item{lifeSpanBuffer:}{numeric scalar; amount by which nominal
lifespan is multiplied to get maximum possible lifespan.}

\item{regStart:}{numeric scalar; if not NULL, search for tags
registered no earlier than this date, and ignore deployment
dates.}

\item{regEnd:}{numeric scalar; if not NULL, search for tags
registered no later than this date, and ignore deployment
dates.}

\item{mfgID:}{character scalar; typically a small integer; return
only records for tags with this manufacturer ID (usually
printed on the tag)}

\item{...:}{additional parameters to motusQuery()}
}
\value{
the list of motus tags and their meta data satisfying the
    search criteria, or NULL if there are none.
}
\description{
return a list of motus tags
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
