% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusUpdateDB.R
\name{motusUpdateDB}
\alias{motusUpdateDB}
\title{update a motus tag detection database (backend)}
\usage{
motusUpdateDB(projRecv, src, countOnly, forceMeta = FALSE)
}
\arguments{
\item{projRecv}{integer scalar project code from motus.org, *or*
character scalar receiver serial number}

\item{src}{src_sqlite object representing the database}

\item{countOnly}{logical scalar: if FALSE, the default, then do
requested database updates.  Otherwise, return a count of items
that would need to be transferred in order to update the
database.}

\item{forceMeta}{logical scalar: if true, re-get metadata for tags and
receivers, even if we already have them.}
}
\value{
\code{src} if \code{countOnly} is FALSE.  Otherwise, a list
of counts of items available for an update.
}
\description{
update a motus tag detection database (backend)
}
\note{
This function does most of the work of fetching data and metadata from
motus servers.  It is not intended to be called directly by users.
}
\seealso{
\code{\link{tagme}}, which is intended for most users, and calls this function.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
