% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusUpdateRecvDB.R
\name{motusUpdateRecvDB}
\alias{motusUpdateRecvDB}
\title{update a motus tag detection database - receiver flavour (backend)}
\usage{
motusUpdateRecvDB(sql, countOnly)
}
\arguments{
\item{sql}{safeSQL object representing the receiver database}

\item{countOnly}{logical scalar: if FALSE, the default, then do
requested database updates.  Otherwise, return a count of items
that would need to be transferred in order to update the
database.}
}
\description{
update a motus tag detection database - receiver flavour (backend)
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
\seealso{
\link{\code{tagme}}, which is intended for most users, and
    indirectly calls this function.
}

