% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusUpdateTagDB.R
\name{motusUpdateTagDB}
\alias{motusUpdateTagDB}
\title{update a motus tag detection database - tag flavour (backend)}
\usage{
motusUpdateTagDB(src, countOnly = FALSE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{countOnly}{logical scalar: count results instead of returning them?}
}
\value{
\code{src}, if countOnly is FALSE.  Otherwise, a list
of counts items that would be transferred by the update.
}
\description{
update a motus tag detection database - tag flavour (backend)
}
\seealso{
\link{\code{tagme}}, which is intended for most users, and
    indirectly calls this function.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
