% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_track}
\alias{plot_track}
\title{Plot track}
\usage{
plot_track(track, title = "", discrete = FALSE)
}
\arguments{
\item{track}{track data.frame from \code{\link{fit_ssm}}}

\item{title}{title of plot}

\item{discrete}{color movement by discrete state? If FALSE, colored by gamma}
}
\description{
Plot track
}
\examples{

}
