% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot temporal trend}
\usage{
plot_trend(track, y_name = "lon", xlab = "time", ylab = y_name,
  col_est = "steelblue", col_obs = "black")
}
\arguments{
\item{track}{track data.frame from \code{\link{fit_ssm}}}

\item{y_name}{name of variable to plot}

\item{xlab}{label for x axis}

\item{ylab}{label for y axis}

\item{col_est}{colour for model estimates}

\item{col_obs}{colour for observations}
}
\description{
Plot temporal trend
}
