% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRouteMap.R
\name{plotRouteMap}
\alias{plotRouteMap}
\title{Map of tag routes and sites coloured by id}
\usage{
plotRouteMap(
  src,
  maptype = "osm",
  zoom = NULL,
  start_date = NULL,
  end_date = NULL,
  lim_lat = NULL,
  lim_lon = NULL,
  data,
  lat,
  lon,
  recvStart,
  recvEnd
)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}

\item{maptype}{Character. Map tiles to use. Must be one of \code{rosm::osm.types()},
such as \code{osm}, \code{stamenbw}, etc. Most map tiles require attribution for
publication, see details.}

\item{zoom}{Integer. Override the calculated zoom level to increase or
decrease the resolution of the map tiles.}

\item{start_date}{Character. Optional start date for routes.}

\item{end_date}{Character. Optional end date for routes.}

\item{lim_lat}{Numeric vector. Optional latitudinal plot limits.}

\item{lim_lon}{Numeric vector. Optional longitudinal plot limits.}

\item{data}{Defunct, use \code{src}, \code{df_src}, or \code{df} instead.}

\item{lat}{Defunct}

\item{lon}{Defunct}

\item{recvStart}{Defunct}

\item{recvEnd}{Defunct}
}
\description{
Google map of routes of Motus tag detections coloured by ID.  User defines a
date range to show points for receivers that were operational at some point
during the date range.
}
\details{
By default this function uses OSM maps (Open Street Map). OSM and
many other map tiles are released under specific licences, which generally
require that you give attribution at a minimum. See
\href{https://www.openstreetmap.org/copyright}{OSM} for more details on their
tiles, but remember to check what other groups require if you use their
tiles.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# Plot route map of all detection data, with "osm" maptype, and receivers
# active between 2016-01-01 and 2017-01-01
plotRouteMap(sql_motus, start_date = "2016-01-01", end_date = "2016-12-31")
\dontshow{\}) # examplesIf}
}
