% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSite.R
\name{plotSite}
\alias{plotSite}
\title{Plot all tags by site}
\usage{
plotSite(data, sitename = unique(data$site))
}
\arguments{
\item{data}{tbl file of .motus data}
}
\description{
Plot tag ID vs time for all tags detected by site, coloured by antenna bearing
}
\examples{
access the "all tags" table within the motus sql
tmp <- tbl(motusSqlFile, "alltags")

Plot all sites within file
plotSite(tmp)

Plot only detections at a specific site; Piskwamish
plotSite(filter(tmp, site == "Piswamish"))

}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
