% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiteSig.R
\name{plotSiteSig}
\alias{plotSiteSig}
\title{Plot signal strength of all tags by a specified site}
\usage{
plotSiteSig(data, recvDepName)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for antBearing, ts, lat, sig, fullID, recvDepName}

\item{recvDepName}{name of recvDepName}
}
\description{
Plot signal strength vs time for all tags detected at a specified site, coloured by antenna
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
