% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRouteMap.R
\name{points2Path}
\alias{points2Path}
\title{Convert points to path}
\usage{
points2Path(df, by = "fullID", lat = "recvDeployLat", lon = "recvDeployLon")
}
\arguments{
\item{df}{Data frame. Could be the result of \code{dplyr::tbl(tags, "alltags") \%>\% dplyr::collect()}.}

\item{by}{Character. Column defining the tag id over which to group points
into paths. Defaults to "fullID".}

\item{lat}{Character. Name of column with latitude values, defaults to
\code{recvDeployLat}.}

\item{lon}{Character. Name of column with longitude values, defaults to
\code{recvDeployLon}.}
}
\value{
Spatial data frame with MULTILINE paths
}
\description{
Converts a data frame with a list of lat/lons to a spatial data frame with
MULTILINES defining paths by tag id. Useful for plotting with
\code{ggplot2::geom_sf()}. Will silently remove single points.
}
