% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeSQL.R
\name{print.safeSQL}
\alias{print.safeSQL}
\title{safeSQL method for printing}
\usage{
\method{print}{safeSQL}(x, ...)
}
\arguments{
\item{x}{safeSQL object}

\item{...}{For print() compatibility}
}
\value{
invisible(NULL)
}
\description{
safeSQL method for printing
}
\keyword{internal}
