% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{samp_ssm}
\alias{samp_ssm}
\title{Draw MCMC samples from state-space model}
\usage{
samp_ssm(model, ...)
}
\arguments{
\item{model}{Object produced by \code{\link{fit_ssm}}}

\item{...}{Additional arguments to pass to \code{\link{tmbstan::tmbstan}}}
}
\description{
Draw MCMC samples from state-space model
}
\details{
Simple wrapper for \code{\link{tmbstan::tmbstan}}
}
