% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqliteToRDS.R
\name{serializeNoHeader}
\alias{serializeNoHeader}
\title{serialize an object to a raw vector, without the "RDS" file header.}
\usage{
serializeNoHeader(x, dropTypeLen = FALSE)
}
\arguments{
\item{x}{the R object}

\item{dropTypeLen}{logical; if TRUE, also drop the initial TYPE and length fields (8 extra bytes)}
}
\value{
raw vector to which x has been serialized in binary little-endian (non-XDR) format,
but without the leading 'B\\n' and three 32-bit integers of header:
   RDS_serialization_version
   RDS_R_writer_version
   RDS_min_R_version
So we just drop the first 14 bytes.
}
\description{
serialize an object to a raw vector, without the "RDS" file header.
}
\note{
this is a convenience function used by sqliteToRds, and is intended for small objects,
since the serialization is done in-memory.
}
