% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sessionVariable.R
\name{sessionVariable}
\alias{sessionVariable}
\title{Create a session variable in an environment.}
\usage{
sessionVariable(name, info = name, env = Motus, class = "character",
  val = NULL)
}
\arguments{
\item{name}{name of variable}

\item{info}{human-readable description of the variable. This is used
to prompt the user if the value of the variable is not already known.

Alternatively, this can be an R function to be called the first time
the value needs to be obtained.  That function will presumably
depend on other session variables.  The function should return
the value of the variable (or generate an error if not successful).}

\item{env}{environment in which to create the session variable
Default:  Motus}

\item{class}{the class of the variable desired; when the user is
prompted for it, the value is coerced to this class.
Default: "character"}

\item{val}{initial value of the session variable.  Default: unset.}
}
\value{
a function which can be installed as an active binding in an
environment
}
\description{
Session variables are typically only set once per R session.
They include items such as user login and password for motus.org
They are implemented as active bindings in an environment.  If the
binding's current value is NULL, the user is prompted to enter
the value.
}
\examples{

 sessionVariable("userLogin", "user login at motus.org")

## This creates an active binding for the symbol "userLogin" in the environment "Motus"
## If the variable \\code{Motus$userLogin} is requested in code, the user will
## be prompted to enter a value if there is no current value.

}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
