% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shorebirds}
\alias{shorebirds}
\title{Example shorebirds tags download from Motus}
\format{A data frame with 108826 rows and 60 variables:}
\usage{
shorebirds
}
\description{
An example data frame of shorebirds data (project 176)
}
\keyword{datasets}
