% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sim_ssm}
\alias{sim_ssm}
\title{Simulate a track using parameters estimated by state-space model}
\usage{
sim_ssm(model)
}
\arguments{
\item{model}{Object produced by \code{\link{fit_ssm}}}
}
\description{
Simulate a track using parameters estimated by state-space model
}
\details{
Essentially a wrapper for \code{model$obj$simulate()}
}
