% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSiteDet.R
\name{simSiteDet}
\alias{simSiteDet}
\title{Create a dataframe of simultaneous detections at multiple sites}
\usage{
simSiteDet(data)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for motusTagID, recvDepName, ts}
}
\description{
Creates a dataframe consisting of detections of tags that are detected at two or more receiver
at the same time.
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
