% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvActivityForBatches.R
\name{srvActivityForBatches}
\alias{srvActivityForBatches}
\title{get the activity of specific batches from the data server}
\usage{
srvActivityForBatches(batchID = 0, ant = NULL, hourBin = NULL)
}
\arguments{
\item{batchID}{Integer. Next batchID to query}

\item{ant}{Integer. Next antenna to query}

\item{hourBin}{Integer. Next hourBin to query}
}
\value{
data.frame with these columns:
\itemize{
\item batchID integer motus batch ID
\item motusDeviceID integer motus device ID
\item monoBN boot number, for SG receivers; NA for Lotek
\item tsStart first timestamp of data
\item tsEnd last timestamp of data
\item numHits integer number of hits
\item ts real processing timestamp
}
}
\description{
get the activity of specific batches from the data server
}
\keyword{internal}
