% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvAPIinfo.R
\name{srvAPIinfo}
\alias{srvAPIinfo}
\title{get information about the motus data server API}
\usage{
srvAPIinfo()
}
\value{
a list with (at least) these items:
\itemize{
   \item maxRows; integer; maximum number of rows returned by any API calls
}
}
\description{
The server is queried for a few details about the API.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
