% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvBatchesForAll.R
\name{srvBatchesForAll}
\alias{srvBatchesForAll}
\title{get new batches for all receivers from the data server}
\usage{
srvBatchesForAll(batchID = 0)
}
\arguments{
\item{batchID}{integer largest batchID already obtained.
Default: 0, meaning none.}
}
\value{
data.frame with these columns:
\itemize{
\item batchID integer motus batch ID
\item motusDeviceID integer motus device ID
\item monoBN boot number, for SG receivers; NA for Lotek
\item tsStart first timestamp of data
\item tsEnd last timestamp of data
\item numHits integer number of hits
\item ts real processing timestamp
}
}
\description{
get new batches for all receivers from the data server
}
\details{
this call is used to obtain batches incrementally, regardless of
which receiver they come from.  It is only available to administrative users,
and caling it as a non-administrative user stops with an error.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
