% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvBatchesForTagProject.R
\name{srvBatchesForTagProject}
\alias{srvBatchesForTagProject}
\title{get the batches for a tag project from the data server}
\usage{
srvBatchesForTagProject(projectID, ts = 0, countOnly = FALSE)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{ts}{real scalar processing timestamp of latest batch already owned
Default: 0, meaning none.}

\item{countOnly}{logical; if TRUE, return only the cound of available batches.
Default: FALSE.}
}
\value{
data.frame with these columns:
\itemize{
\item batchID integer motus batch ID
\item deviceID integer motus device ID
\item monoBN boot number, for SG receivers; NA for Lotek
\item tsBegin first timestamp of data
\item tsEnd last timestamp of data
\item numHits integer number of hits
\item ts real processing timestamp
}
}
\description{
get the batches for a tag project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

