% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvDeviceIdForReceiver.R
\name{srvDeviceIDForReceiver}
\alias{srvDeviceIDForReceiver}
\title{get the device ID for one or more receivers}
\usage{
srvDeviceIDForReceiver(serno)
}
\arguments{
\item{serno}{character vector of receiver serial numbers, e.g. "SG-1234BBBK4321", "Lotek-123"}
}
\value{
a data.frame with these columns:
\itemize{
\item serno; character serial number, e.g. "SG-1214BBBK3999", "Lotek-8681"
\item deviceID; integer device ID (internal to motus); NA if the serial number
was not valid or not known.
}
}
\description{
The deviceID is returned for any valid receiver serial numbers.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
