% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvGPSforTagProject.R
\name{srvGPSforTagProject}
\alias{srvGPSforTagProject}
\title{get the GPS fixes for a tag project from the data server}
\usage{
srvGPSforTagProject(projectID, batchID, ts = 0)
}
\arguments{
\item{projectID}{integer scalar motus tag project ID}

\item{batchID}{integer scalar batch ID}

\item{ts}{real scalar processing timestamp of latest fix already owned
Default: 0, meaning none.}
}
\value{
data.frame with these columns:
\itemize{
   \item ts       numeric system timestamp
   \item gpsts    numeric GPS timestamp
   \item batchID  integer batch ID
   \item lat      numeric latitude in degrees N (negative is south)
   \item lon      numeric longitude in degrees E (negative is west)
   \item alt      numeric altitude in metres ASL
}
}
\description{
These are the periodic GPS fixes from receivers that detected
tags from the project.
}
\note{
: this returns fixes from receivers which detected tags from the
given project.  Fixes from up to 1 hour before to 1 hour after the
detections are included, to ensure temporal coverage.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
