% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvHitsForTagProject.R
\name{srvHitsForTagProject}
\alias{srvHitsForTagProject}
\title{get the hits for a tag project from the data server}
\usage{
srvHitsForTagProject(projectID, batchID, hitID = 0)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{batchID}{integer scalar motus batch ID}

\item{hitID}{integer scalar ID of latest hit already obtained.
Default: 0, meaning none.}
}
\value{
data.frame with these columns:
\itemize{
  \item hitID
  \item runID
  \item batchID
  \item ts
  \item sig
  \item sigSD
  \item noise
  \item freq
  \item freqSD
  \item slop
  \item burstSlop
}
}
\description{
get the hits for a tag project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
