% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvProjectAmbiguitiesForTagProject.R
\name{srvProjectAmbiguitiesForTagProject}
\alias{srvProjectAmbiguitiesForTagProject}
\title{get the project ambiguities for a tag project from the data server}
\usage{
srvProjectAmbiguitiesForTagProject(projectID)
}
\arguments{
\item{projectID}{integer scalar motus project ID}
}
\value{
data.frame with these columns:
\itemize{
\item ambigProjectID integer negative project ambiguity ID
which represents a unique set of project IDs sharing at least
one ambiguous tag detection
\item projectID1 motus project ID
\item projectID2 motus project ID
\item projectID3 motus project ID
\item projectID4 motus project ID
\item projectID5 motus project ID
\item projectID6 motus project ID
}
}
\description{
get the project ambiguities for a tag project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
