% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvPulseCountsForReceiver.R
\name{srvPulseCountsforReceiver}
\alias{srvPulseCountsforReceiver}
\title{get the pulse counts for an antenna from the data server}
\usage{
srvPulseCountsforReceiver(batchID, ant, hourBin = 0)
}
\arguments{
\item{batchID}{integer scalar batch ID}

\item{ant}{integer antenna}

\item{hourBin}{hour bin, i.e. \code{floor(timestamp/3600)} for the
pulses in this bin.  Default: 0, meaning no pulse counts have been
obtained for this batch.  When 0, \code{ant} is ignored}
}
\value{
data.frame with these columns:
\itemize{
   \item batchID  integer batch ID
   \item ant      integer antenna number
   \item hourBin  numeric hour indicator given as \code{floor(timestamp / 3600)}
   \item count    integer number of pulses seen by antenna in this hour bin
}
}
\description{
Pulse counts summarize antenna activity on an hourly basis.
}
\note{
Paging for this query is handled by passing the last ant and hourBin values
received to the next call of this function.  Values are returned sorted by
\code{hourBin} \emph{within} \code{ant} for each batchID.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
