% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvQuery.R
\name{srvQuery}
\alias{srvQuery}
\title{Send a query to the data server API}
\usage{
srvQuery(API, params = NULL, requestType = "post", show = FALSE,
  JSON = FALSE, auth = TRUE)
}
\arguments{
\item{API}{one of the Motus$API_DATA_... constants}

\item{params}{named list of API-specific parameters}

\item{requestType}{"post" or "get"; default: "post".}

\item{show}{if TRUE, print the request to the console before submitting to motus}

\item{JSON}{if TRUE, return results as JSON-format string; otherwise, as R list}

\item{auth}{if TRUE, the default, the API requires an
    authentication token, which will be included automatically from \code{Motus$authToken}
Otherwise, an authentication token will be requested using the credentials in \code{Motus[c("userLogin", "userPassword")]}}
}
\value{
the result of sending the request to the data server API.  The
    result is a JSON-format character scalar if \code{json} is
    \code{TRUE}; otherwise it is an R list with named components,
    extracted from the JSON return value.
}
\description{
Send a query to the data server API
}
\note{
If you have not already entered your motus login credentials
    in this R session, you will be prompted for them.
}
\author{
John Brzustowski
    \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
