% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvRunsForReceiver.R
\name{srvRunsForReceiver}
\alias{srvRunsForReceiver}
\title{get the runs for a receiver from the data server}
\usage{
srvRunsForReceiver(batchID, runID = 0)
}
\arguments{
\item{batchID}{integer scalar motus batch ID}

\item{runID}{integer scalar ID of latest run already obtained.
Default: 0, meaning none.}
}
\value{
a data.frame with these columns:
\itemize{
   \item runID
   \item batchIDbegin
   \item tsBegin
   \item tsEnd
   \item done
   \item motusTagID
   \item ant
   \item len
}
}
\description{
get the runs for a receiver from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
