% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvRunsForReceiverProject.R
\name{srvRunsForReceiverProject}
\alias{srvRunsForReceiverProject}
\title{get the runs for a receiver project from the data server}
\usage{
srvRunsForReceiverProject(projectID, batchID, runID = 0)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{batchID}{integer scalar motus batch ID}

\item{runID}{integer scalar ID of latest run already obtained.
Default: 0, meaning none.}
}
\value{
a list with these items:
\itemize{
   \item runs, a data.frame with these columns:
      \itemize{
         \item runID
         \item batchIDbegin
         \item tsBegin
         \item tsEnd
         \item done
         \item motusTagID
         \item ant
         \item len
      }
   \item batchRuns, a data.frame with these columns:
      \itemize{
         \item batchID
         \item runID
      }
}
}
\description{
get the runs for a receiver project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
