% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvSizeOfUpdateForTagProject.R
\name{srvSizeOfUpdateForTagProject}
\alias{srvSizeOfUpdateForTagProject}
\title{get the amount of data required to update a tag project from the data server}
\usage{
srvSizeOfUpdateForTagProject(projectID, batchID = 0)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{batchID}{integer largest batchID already owned for this project.
Default: 0, meaning none.}
}
\value{
a list with these items:
\itemize{
\item numBatches - number of new batches
\item numRuns - number of new or updated runs
\item numHits - number of new hits
\item numGPS - number of GPS fixes
\item size - estimate of transfer size, in bytes
}
}
\description{
get the amount of data required to update a tag project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
