% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvTagsForAmbiguities.R
\name{srvTagsForAmbiguities}
\alias{srvTagsForAmbiguities}
\title{get the motus tagIDs for amibiguity IDs}
\usage{
srvTagsForAmbiguities(ambigIDs)
}
\arguments{
\item{ambigIDs}{integer vector of negative ambiguity IDs}
}
\value{
a data.frame with these columns:
\itemize{
   \item ambigID; negative integer tag ambiguity ID
   \item motusTagID1; positive integer motus tag ID
   \item motusTagID2; positive integer motus tag ID
   \item motusTagID3; positive integer motus tag ID or NA
   \item motusTagID4; positive integer motus tag ID or NA
   \item motusTagID5; positive integer motus tag ID or NA
   \item motusTagID6; positive integer motus tag ID or NA
}
with one row per ambigID
}
\description{
Ambiguity IDs ("negative tag IDs") represent sets of 2 to 6 motus tags
which cannot be distinguished over some period of time.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
