% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tagmeSample}
\alias{tagmeSample}
\title{Create an in-memory copy of sample tags data}
\usage{
tagmeSample(db = "project-176.motus")
}
\arguments{
\item{db}{Character. Name of sample data base to load. The sample data is
"project-176.motus".}
}
\description{
For running examples and testing out motus functionality, it can be useful to
work with sample data set. You can download the most up-to-date copy of this
data yourself (to \verb{project-176.motus}) with the username and password both
"motus.sample".
}
\details{
\code{sql_motus <- tagme(176, new = TRUE, update = TRUE)}

Or you can use this helper function to grab an in-memory copy bundled in this
package.
}
\examples{
# Explore the sample data
tags <- tagmeSample()
dplyr::tbl(tags, "activity")
dplyr::tbl(tags, "alltags")
}
