% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagSum.R
\name{tagSum}
\alias{tagSum}
\title{General summary of detections for each tag}
\usage{
tagSum(df_src, data)
}
\arguments{
\item{df_src}{Data frame, SQLite connection, or SQLite table. An SQLite
connection would be the result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}; an SQLite table would be
the result of \code{dplyr::tbl(tags, "alltags")}; a data frame could be the
result of \code{dplyr::tbl(tags, "alltags") \%>\% dplyr::collect()}.}

\item{data}{Defunct, use \code{src}, \code{df_src}, or \code{df} instead.}
}
\value{
A flat data frame with the following for each tag:
\itemize{
\item \code{fullID} - \code{fullID} of Motus registered tag
\item \code{first_ts} - Time (\code{ts}) of first detection
\item \code{last_ts} - Time (\code{ts}) of last detection
\item \code{first_site} - First detection site (\code{recvDeployName})
\item \code{last_site} - Last detection site (\code{recvDeployName})
\item \code{recvLat.x} - Latitude of first detection site (\code{gpsLat} or \code{recvDeployLat})
\item \code{recvLon.x} - Longitude of first detection site (\code{gpsLon} or \code{recvDeployLon})
\item \code{recvLat.y} - Latitude of last detection site (\code{gpsLat} or \code{recvDeployLat})
\item \code{recvLon.y} - Longitude of last detection site (\code{gpsLon} or \code{recvDeployLon})
\item \code{tot_ts} - Time between first and last detection (in seconds)
\item \code{dist} - Straight line distance between first and last detection site (in metres)
\item \code{rate} - Overall rate of movement (\code{tot_ts}/\code{dist}), in metres/second
\item \code{bearing} - Bearing between first and last detection sites
\item \code{num_det} - Number of detections summarized
}
}
\description{
Creates a summary for each tag of it's first and last detection time (\code{ts}),
first and last detection site, length of time between first and last
detection, straight line distance between first and last detection site, rate
of movement, and bearing. Lat/lons are taken from \code{gpsLat}/\code{gpsLon}, or if
missing, from \code{recvDeployLat}/\code{recvDeployLon}. Bearing is calculated using
the \code{geosphere::bearing()} function.
}
\examples{

# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# Summarize tags
tag_summary <- tagSum(sql_motus)

# For specific SQLite table/view (needs gpsLat/gpsLon) --------------
library(dplyr)
tbl_alltagsGPS <- tbl(sql_motus, "alltagsGPS") 
tag_summary <- tagSum(tbl_alltagsGPS)

# For a flattened data frame ----------------------------------------
df_alltagsGPS <- collect(tbl_alltagsGPS)
tag_summary <- tagSum(df_alltagsGPS)

# Can be filtered, e.g., for only a few tags
tag_summary <- tagSum(filter(tbl_alltagsGPS, motusTagID \%in\% c(16047, 16037, 16039)))
}
