% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tellme.R
\name{tellme}
\alias{tellme}
\title{Report how much new data motus has for a tag detection database}
\usage{
tellme(projRecv, new = FALSE, dir = getwd())
}
\arguments{
\item{projRecv}{Numeric. Project code from motus.org, \emph{or} character receiver
serial number.}

\item{new}{Logical. Create a new database (Default \code{FALSE})? Specify
\code{new = TRUE} to create a new local copy of the database to be downloaded.
Otherwise, it assumes the database already exists, and will stop with an
error if it cannot find it in the current directory. This is mainly to
prevent inadvertent downloads of large amounts of data that you already
have!}

\item{dir}{Character. Path to the folder where you are storing databases
IF \code{NULL} (default), uses current working directory.}
}
\value{
a named list with these items:
\itemize{
\item numBatches: number of batches having data for your database
\item numRuns: number of runs of tags detections with new data
\item numHits: number of new detections
\item numGPS: number of new GPS fixes covering the new detections
\item numBytes: estimated size of download, in bytes.  This is an estimate of the
\emph{uncompressed} size, but data are gz-compressed for transfer, so the number
of bytes you have to download is typically going to be smaller than this
number by a factor of 2 or more.
}
}
\description{
"new" means data not already in your local database.
}
\note{
if you specify \code{new = TRUE} and the database does not already exist, it
will be created (but empty).
}
