% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMotusDb.r
\name{updateMotusDb}
\alias{updateMotusDb}
\title{Ensures that the motus sqlite file is up-to-date to support the current version of the package.}
\usage{
updateMotusDb(rv, src, projRecv, deviceID)
}
\arguments{
\item{rv}{return value}

\item{src}{sqlite database source}

\item{projRecv}{parameter provided by the hook function call, when opening a file built by project ID}

\item{deviceID}{parameter provided by the hook function call, when opening a file built by receiver ID}
}
\value{
rv
}
\description{
This function is called from the z.onLoad function which adds a hook to the ensureDBTables function of the motusClient package.
addHook("ensureDBTables", updateMotusDb). I.E., the current function will be called each time that a new motus file is opened
(and the ensureDBTables function is accessed).
}
\author{
Denis Lepage \email{dlepage@bsc-eoc,org}
}
