% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRunsFilter.R
\name{writeRunsFilter}
\alias{writeRunsFilter}
\title{Write to the local database the probabilities associated with runs for a filter}
\usage{
writeRunsFilter(
  src,
  filterName,
  motusProjID = NA,
  df,
  overwrite = TRUE,
  delete = FALSE
)
}
\arguments{
\item{src}{SQLite connection (result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")})}

\item{filterName}{Character. Unique name given to the filter}

\item{motusProjID}{Character. Optional project ID attached to the filter in
order to share with other users of the same project.}

\item{df}{Data frame. Containing \code{runID}, \code{motusTagID} and probability values
to save in the local database}

\item{overwrite}{Logical. When \code{TRUE} ensures that existing records matching
the same \code{filterName} and \code{runID} get replaced}

\item{delete}{Logical. When TRUE, removes all existing filter records
associated with the \code{filterName} and re-inserts the ones contained in the
dataframe. This option should be used if the dataframe provided contains
the entire set of filters you want to save.}
}
\value{
database connection refering to the filter created
}
\description{
Write to the local database the probabilities associated with runs for a filter
}
