% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Add/update batch activity}
\usage{
activity(src, resume = FALSE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}}

\item{resume}{Logical. Resume a download? Otherwise the table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the \code{activity} table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\details{
This function is automatically run by the \code{\link[=tagme]{tagme()}} function with
\code{resume = TRUE}.

If an \code{activity} table doesn't exist, it will be created prior to
downloading. If there is an existing \code{activity} table, this will update the
records.
}
\examples{

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
  
# Access 'activity' table
library(dplyr)
a <- tbl(sql.motus, "activity")
  
# If interrupted and you want to resume
\dontrun{my_tags <- activity(sql.motus, resume = TRUE)}

}
