% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Add/update batch activity}
\usage{
activity(src, resume = FALSE)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}

\item{resume}{Logical. Resume a download? Otherwise the table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the \code{activity} table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\details{
This function is automatically run by the \code{\link[=tagme]{tagme()}} function with
\code{resume = TRUE}.

If an \code{activity} table doesn't exist, it will be created prior to
downloading. If there is an existing \code{activity} table, this will update the
records.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE, update = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()
   
# Access 'activity' table
library(dplyr)
a <- tbl(sql_motus, "activity")
  
# If interrupted and you want to resume
\dontrun{my_tags <- activity(sql_motus, resume = TRUE)}

}
