% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkVersion.R
\name{checkVersion}
\alias{checkVersion}
\title{Verifies the version of the package against the admInfo table of a motus file. Those should match
if the updateMotusDb function has been properly loaded by the tagme function.}
\usage{
checkVersion(src)
}
\arguments{
\item{src}{motus sqlite database source}
}
\description{
Verifies the version of the package against the admInfo table of a motus file. Those should match
if the updateMotusDb function has been properly loaded by the tagme function.
}
