% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteRunsFilter.R
\name{deleteRunsFilter}
\alias{deleteRunsFilter}
\title{Delete a filter  matching a filter name (and optionally a project ID)}
\usage{
deleteRunsFilter(src, filterName, motusProjID = NA, clearOnly = FALSE)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{filterName}{unique name given to the filter}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{clearOnly}{boolean. When true, only remove the probability records associated with the filter,
but retain the filter itself}
}
\value{
the integer filterID of the filter deleted
}
\description{
Delete a filter  matching a filter name (and optionally a project ID)
}
