% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecateBatches.R
\name{deprecateBatches}
\alias{deprecateBatches}
\title{Fetch and remove deprecated batches}
\usage{
deprecateBatches(src, fetchOnly = FALSE, ask = TRUE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{fetchOnly}{Logical. Only \emph{fetch} batches that are deprecated. Don't
remove deprecated batches from other tables.}

\item{ask}{Logical. Ask for confirmation when removing deprecated batches}
}
\description{
Deprecated batches are removed from the online database but not from
local data files. This function fetches a list of deprecated batches
(stored in the 'deprecated' table), and, optionally, removes these batches
from all tables that reference \code{batchID}s
}
