% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensureDBTables.R
\name{ensureDBTables}
\alias{ensureDBTables}
\title{make sure a tag or receiver database has the required tables}
\usage{
ensureDBTables(src, projRecv, deviceID)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{projRecv}{integer scalar motus project ID number *or*
    character scalar receiver serial number; must be specified if
\code{src} does not already contain a table named \code{meta}.}

\item{deviceID}{integer scalar motus deviceID; must be specified
when this is a new receiver database.}
}
\value{
returns a dplyr::tbl representing the alltags virtual table
which is created in \code{src}.
}
\description{
make sure a tag or receiver database has the required tables
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
