% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_ssm}
\alias{fit_ssm}
\title{Function for fitting state-space movement model}
\usage{
fit_ssm(track, formula = NULL, gamma_model = "RW",
  gamma_threshold = 0.8, gamma_prob = 0.95, dist = "t",
  silent = FALSE, gr_threshold = 10, start_par = NULL,
  scale = "sd")
}
\arguments{
\item{track}{data.frame of an individual track containing time, UTM coordinates, time differences
and estimates of observation error. UTM coordinate columns should be labeled
lon and lat and units should equal km. time difference should be a ratio of
the minimum sampling interval and named delta_t. Observation error should be labeled
1) sd_lon and sd_lat for the normal distribution; 2) tau_lon, tau_lat, nu_lon,
and nu_lat for the parameters for the t-distribution; and, 3) scale_lon and
scale_lat for the cauchy distribution.}

\item{formula}{Formula describing relationship between gamma (parameter controlling autocorrelation
in the movement process) and covariates. Covariates are not used if set to NULL}

\item{gamma_model}{Should the gamma parameter (autocorrelation for movement) be "fixed", modeled
as a random walk ("RW") or an "AR1" process?}

\item{gamma_threshold}{A threshold for defining directed or area-restricted phases of movement}

\item{dist}{Distribution to use for observation error ("null", "normal", "t", "cauchy", "robust").
The "robust" option is a mixture of the normal and t distributions.
If NULL, locations are to be "true".}

\item{silent}{Disable tracing information?}

\item{gr_threshold}{Stop if maximum gradient exceeds this value (large values indicate convergence issues)}

\item{start_par}{List of start parameters from a simpler fit from this model (par_est in the list).
It may be helpful to fit a simpler version of this model, then supply starting parameters
that are closer to where they should be.}

\item{scale}{Method for scaling x values to improve parameter scaling and aid convergence, where
"sd" scales using sd, "max" scales to max deviation, or provide a value.}

\item{gamma_prop}{Probability of being considered above or below gamma_threshold value
(i.e. "directed" or "area-restricted")}
}
\description{
Function for fitting state-space movement model
}
