% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_ssm}
\alias{fit_ssm}
\title{Function for fitting state-space movement model}
\usage{
fit_ssm(track, scale = "sd", fix_gamma = FALSE, dist = "t",
  silent = FALSE, gr_threshold = 10)
}
\arguments{
\item{track}{data.frame of an individual track containing time, UTM coordinates, time differences
and estimates of observation error. UTM coordinate columns should be labeled
lon and lat and units should equal km. time difference should be a ratio of
the minimum sampling interval and named delta_t. Observation error should be labeled
1) sd_lon and sd_lat for the normal distribution; 2) tau_lon, tau_lat, nu_lon,
and nu_lat for the parameters for the t-distribution; and, 3) scale_lon and
scale_lat for the cauchy distribution.}

\item{scale}{Method for scaling x values to improve parameter scaling and aid convergence, where
"sd" scales using sd, "max" scales to max deviation, or provide a value.}

\item{fix_gamma}{Should the gamma parameter (autocorrelation for movement) be fixed or time-varrying?}

\item{dist}{Distribution to use for observation error ("null", "normal", "t", or "cauchy").
If "null", locations are to be "true".}

\item{silent}{Disable tracing information?}

\item{gr_threshold}{Stop if maximum gradient exceeds this value (large values indicate convergence issues)}
}
\description{
Function for fitting state-space movement model
}
