% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunsFilters.R
\name{getRunsFilters}
\alias{getRunsFilters}
\title{Get runsFilters}
\usage{
getRunsFilters(src, filterName, motusProjID = NA)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}

\item{filterName}{Character. Unique name given to the filter}

\item{motusProjID}{Character. Optional project ID attached to the filter in
order to share with other users of the same project.}
}
\value{
a database connection to \code{src}
}
\description{
Returns a dataframe of the \code{runsFilters} records matching a filter name (and
optionally a project ID) stored in the local database.
}
