% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsAll.R
\name{gpsAll}
\alias{gpsAll}
\title{Add/update all GPS points}
\usage{
gpsAll(src, resume = TRUE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{resume}{Logical. Resume a download? Otherwise the \code{activity} table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the \code{gpsAll} table in an existing Motus
database. Batch activity refers to the number of hits detected during a given
batch. Batches with large numbers of hits may indicate interference and thus
unreliable hits.
}
\details{
This function is automatically run by the \code{\link[=tagme]{tagme()}} function with
\code{resume = TRUE}.

If an \code{activity} table doesn't exist, it will be created prior to
downloading. If there is an existing \code{activity} table, this will update the
records.
}
\examples{

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
  
# Get all GPS points
\dontrun{sql.motus <- gpsAll(sql.motus)}

# Access 'gpsAll' table
library(dplyr)
g <- tbl(sql.motus, "gpsAll")
  
# gpsAll resumes a previous download by default
# If you want to delete this original data and do a fresh download, 
# use resume = FALSE
\dontrun{sql.motus <- gpsAll(sql.motus, resume = FALSE)}

}
