% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitsBlu.R
\name{hitsBlu}
\alias{hitsBlu}
\title{Update hitsBlu data}
\usage{
hitsBlu(src)
}
\arguments{
\item{src}{SQLite connection. Result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}.}
}
\description{
Add any missing BLU tag hits to the \code{hitsBlu} table in an existing Motus
database. \code{hitsBlu} contain extra information regarding the 'health' of
portable node units. Use \code{\link[=getBluPayload]{getBluPayload()}} to extract these details from
the payload data.
}
\details{
This function is only required if you suspect BLU tag hits have been
missed (due to hits being downloaded before the motus package had the
functionality to download BLU tag hits).
}
\examples{
\dontrun{
  hitsBlu(my_tags)
}

}
\seealso{
\code{\link[=getBluPayload]{getBluPayload()}}
}
