% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{update entire metadata for receivers and tags from Motus server. Contrary to tagme, 
this function retrieves the entire set of metadata for tags and receivers, and not only those 
pertinent to the detections of your local file.}
\usage{
metadata(src, projectIDs = NULL, replace = TRUE, delete = FALSE)
}
\arguments{
\item{src}{src_sqlite object representing the database (either tag or receiver)}

\item{projectIDs}{optional integer vector of Motus projects IDs for which metadata should
be obtained; default: NULL, meaning obtain metadata for all tags and receivers that your permissions 
allow.}

\item{replace}{logical scalar; if TRUE (default), existing data replace the existing metadata with the newly acquired ones.}

\item{delete}{logical scalar; Defaut = FALSE. if TRUE, the entire metadata tables are cleared (for all projects) before re-importing the metadata.}
}
\description{
update entire metadata for receivers and tags from Motus server. Contrary to tagme, 
this function retrieves the entire set of metadata for tags and receivers, and not only those 
pertinent to the detections of your local file.
}
\seealso{
\code{\link{tagme}} provides an option to update only the metadata relevant to a specific 
project or receiver file.
}
\author{
Denis Lepage, Bird Studies Canada
}
