% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{update entire metadata for receivers and tags from Motus server. Contrary to tagme,
this function retrieves the entire set of metadata for tags and receivers, and not only those
pertinent to the detections of your local file.}
\usage{
metadata(src, projectIDs = NULL, replace = TRUE, delete = FALSE)
}
\arguments{
\item{src}{src_sqlite object representing the database (either tag or receiver)}

\item{projectIDs}{optional integer vector of Motus projects IDs for which metadata should
be obtained; default: NULL, meaning obtain metadata for all tags and receivers that your permissions
allow.}

\item{replace}{logical scalar; if TRUE (default), existing data replace the existing metadata with the newly acquired ones.}

\item{delete}{logical scalar; Default = FALSE. if TRUE, the entire metadata tables are cleared (for all projects) before re-importing the metadata.}
}
\description{
update entire metadata for receivers and tags from Motus server. Contrary to tagme,
this function retrieves the entire set of metadata for tags and receivers, and not only those
pertinent to the detections of your local file.
}
\examples{

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
                   
# Add extended metadata to your file
\dontrun{metadata(sql.motus)}
  
# Access different metadata tables
library(dplyr)
tbl(sql.motus, "species")
tbl(sql.motus, "projs")
tbl(sql.motus, "tagDeps")
# Etc.
  
}
\seealso{
\code{\link{tagme}} provides an option to update only the metadata relevant to a specific
project or receiver file.
}
