% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motus-pkg.R
\docType{package}
\name{motus}
\alias{motus}
\title{Fetch and use data from the Motus Wildlife Tracking System}
\description{
\code{motus} is an R package for retrieving telemetry data from the Motus
Wildlife Tracking System \url{http://motus.org}.
}
\details{
For a detailed walk-though and instructions check out the
\href{https://motus.org/MotusRBook/}{Motus R Book}

Commonly used functions:
\enumerate{
\item Download telemetry data
\itemize{
\item \code{\link[=tagme]{tagme()}}
\item \code{\link[=tellme]{tellme()}}
\item \code{\link[=metadata]{metadata()}}
\item \code{\link[=checkVersion]{checkVersion()}}
}
\item Create data filters
\itemize{
\item \code{\link[=listRunsFilters]{listRunsFilters()}}
\item \code{\link[=getRunsFilters]{getRunsFilters()}}
\item \code{\link[=createRunsFilter]{createRunsFilter()}}
\item \code{\link[=writeRunsFilter]{writeRunsFilter()}}
\item \code{\link[=deleteRunsFilter]{deleteRunsFilter()}}
}
\item Summarize data
\itemize{
\item \code{\link[=tagSum]{tagSum()}}
\item \code{\link[=tagSumSite]{tagSumSite()}}
\item \code{\link[=simSiteDet]{simSiteDet()}}
\item \code{\link[=siteSum]{siteSum()}}
\item \code{\link[=siteSumDaily]{siteSumDaily()}}
\item \code{\link[=siteTrans]{siteTrans()}}
}
\item Plot data
\itemize{
\item \code{\link[=plotAllTagsCoord]{plotAllTagsCoord()}}
\item \code{\link[=plotAllTagsSite]{plotAllTagsSite()}}
\item \code{\link[=plotDailySiteSum]{plotDailySiteSum()}}
\item \code{\link[=plotRouteMap]{plotRouteMap()}}
\item \code{\link[=plotSite]{plotSite()}}
\item \code{\link[=plotSiteSig]{plotSiteSig()}}
\item \code{\link[=plotTagSig]{plotTagSig()}}
}
\item Sunrises and sets
\itemize{
\item \code{\link[=sunRiseSet]{sunRiseSet()}}
\item \code{\link[=timeToSunriset]{timeToSunriset()}}
}
}

We also include a practice data set:
\itemize{
\item \link{shorebirds}
}
}
\references{
Motus Wildlife Tracking System \url{http://motus.org}
}
