% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusUpdateRecvDB.R
\name{motusUpdateRecvDB}
\alias{motusUpdateRecvDB}
\title{update a motus tag detection database - receiver flavour (backend)}
\usage{
motusUpdateRecvDB(src, countOnly, forceMeta = FALSE)
}
\arguments{
\item{src}{src_sqlite object representing the database}

\item{countOnly}{logical scalar: count results instead of returning them?}

\item{forceMeta}{logical scalar: if true, re-get metadata for tags and
receivers, even if we already have them.  Default: FALSE}
}
\value{
\code{src}, if countOnly is FALSE.  Otherwise, a list of
    counts items that would be transferred by the update.
}
\description{
update a motus tag detection database - receiver flavour (backend)
}
\seealso{
\code{\link{tagme}}, which is intended for most users, and
    indirectly calls this function.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
