% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motusUpdateTagDB.R
\name{motusUpdateTagDB}
\alias{motusUpdateTagDB}
\title{update a motus tag detection database - tag flavour (backend)}
\usage{
motusUpdateTagDB(sql, countOnly = FALSE)
}
\arguments{
\item{sql}{safeSQL object representing the tag project database}

\item{countOnly}{logical scalar: if FALSE, the default, then do
requested database updates.  Otherwise, return a count of items
that would need to be transferred in order to update the
database.}
}
\description{
update a motus tag detection database - tag flavour (backend)
}
\seealso{
\link{\code{tagme}}, which is intended for most users, and
    indirectly calls this function.
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
