% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeData.R
\name{nodeData}
\alias{nodeData}
\title{Add/update nodeData}
\usage{
nodeData(src, resume = FALSE)
}
\arguments{
\item{src}{SQLite connection (result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")})}

\item{resume}{Logical. Resume a download? Otherwise the \code{nodeData} table is
removed and the download is started from the beginning.}
}
\description{
Download or resume a download of the 'nodeData' table in an existing Motus
database. \code{nodeData} contains information regarding the 'health' of portable
node units.
}
\details{
This function is automatically run by the \code{\link[=tagme]{tagme()}} function with
\code{resume = TRUE}.

If an \code{nodeData} table doesn't exist, it will be created prior to
downloading. If there is an existing \code{nodeData} table, this will update the
records.

Note that only records for CTT tags will have the possibility of
\code{nodeData}.

Node metadata is found in the \code{nodeDeps} table, updated along with other
metadata.
}
\examples{

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))
  
# Access `nodeData` table
library(dplyr)
a <- tbl(sql.motus, "nodeData")
  
# If you just want to download `nodeData`
\dontrun{my_tags <- nodeData(sql.motus)}

}
