% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTagsCoord.R
\name{plotAllTagsCoord}
\alias{plotAllTagsCoord}
\title{Plot all tag detections by latitude or longitude}
\usage{
plotAllTagsCoord(data, coordinate = "recvDeployLat", ts = "ts",
  tagsPerPanel = 5)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for recvDepName, fullID, mfgID, date/time, latitude or longitude}

\item{coordinate}{column name from which to obtain location values, by default it is set to recvDeployLat}

\item{ts}{column for a date/time object as numeric or POSIXct, defaults to ts}

\item{tagsPerPanel}{number of tags in each panel of the plot, by default this is 5}
}
\description{
Plot latitude/longitude vs time (UTC rounded to the hour) for each tag using .motus detection data.  
Coordinate is by default taken from a receivers GPS latitude recordings.
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
