% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTagsLat.R
\name{plotAllTagsLat}
\alias{plotAllTagsLat}
\title{Plot all tag detections by latitude}
\usage{
plotAllTagsLat(data, tagsPerPanel = n)
}
\arguments{
\item{data}{dataframe of Motus detection data}

\item{tagsPerPanel}{number of tags in each panel of the plot}
}
\description{
Plot latitude vs time for each tag
}
\examples{
plotAllTagsLat(dat, tagsPerPanel = 4)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
