% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTagsSite.R
\name{plotAllTagsSite}
\alias{plotAllTagsSite}
\title{Plot all tag detections by deployment}
\usage{
plotAllTagsSite(data, coordinate = "recvDeployLat", tagsPerPanel = 5)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for recvDepName, fullID, mfgID, date/time, latitude or longitude}

\item{coordinate}{column of receiver latitude/longitude values to use, defaults to recvDeployLat}

\item{tagsPerPanel}{number of tags in each panel of the plot, default is 5}
}
\description{
Plot deployment (ordered by latitude) vs time (UTC) for each tag
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
