% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAllTagsSite.R
\name{plotAllTagsSite}
\alias{plotAllTagsSite}
\title{Plot all tag detections by site}
\usage{
plotAllTagsSite(data, tagsPerPanel = n)
}
\arguments{
\item{data}{dataframe of Motus detection data}

\item{tagsPerPanel}{number of tags in each panel of the plot}
}
\description{
Plot site (ordered by latitude) vs time for each tag
}
\examples{
plotAllTagsSite(dat, tagsPerPanel = 4)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
