% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDailySiteSum.R
\name{plotDailySiteSum}
\alias{plotDailySiteSum}
\title{Plots number of detections and tags, daily, for a specified site}
\usage{
plotDailySiteSum(data, recvDepName)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for motusTagID, sig, recvDepName, ts}

\item{recvDepName}{name of site to plot}
}
\description{
Plots total number of detections across all tags, and total number of tags detected per day for
a specified site.  Depends on siteSumDaily function.
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
