% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDailySiteSum.R
\name{plotDailySiteSum}
\alias{plotDailySiteSum}
\title{Plots number of detections and tags, daily, for a specified site}
\usage{
plotDailySiteSum(data, Site)
}
\arguments{
\item{data}{dataframe of Motus detection data}

\item{Site}{name of site to plot}
}
\description{
Plots total number of detections across all tags, and total number of tags detected per day for
a specified site.  Depends on siteSumDaily function.
}
\examples{
plotDailySiteSum(dat, Site = "Longridge")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
