% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRouteMap.R
\name{plotRouteMap}
\alias{plotRouteMap}
\title{Map of routes and sites coloured by id}
\usage{
plotRouteMap(site_data, detection_data, zoom, latCentre, lonCentre,
  maptype = c("satellite", "terrain", "hybrid", "roadmap"), startTime,
  endTime)
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\description{
Google map of routes Motus detections coloured by ID, with sites that were operational
at some point during a user defined time interval
}
\examples{
plotRouteMap(site_data = locs, detection_data = dat, maptype = "terrain",
latCentre = 44, lonCentre = -70, zoom = 5, startTime = "2016-01-01", endTime = "2016-12-31")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
