% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSite.R
\name{plotSite}
\alias{plotSite}
\title{Plot all tags by site}
\usage{
plotSite(data, sitename = unique(data$recvDeployName))
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltagsGPS", or a
data.frame of detection data including at a minimum variables for ts,
antBearing, fullID, recvDeployName}

\item{sitename}{Character vector. Which sites to plot? Defaults to all unique
sites.}
}
\description{
Plot tag ID vs time for all tags detected by site, coloured by antenna
bearing
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE, update = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# convert sql file "sql_motus" to a tbl called "tbl_alltags"
library(dplyr)
tbl_alltags <- tbl(sql_motus, "alltagsGPS") 

# convert the tbl "tbl_alltags" to a data.frame called "df_alltags"
df_alltags <- tbl_alltags \%>\% 
  collect() \%>\% 
  as.data.frame()

# Plot all sites within file for tbl file tbl_alltags
plotSite(tbl_alltags)

# Plot only detections at a specific site; Piskwamish for data.frame
# df_alltags
plotSite(filter(df_alltags, recvDeployName == "Piskwamish"))

#Plot only detections for specified tags for data.frame df_alltags
plotSite(filter(df_alltags, motusTagID \%in\% c(16047, 16037, 16039)))

}
