% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSite.R
\name{plotSite}
\alias{plotSite}
\title{Plot all tags by site}
\usage{
plotSite(data, sitename = unique(data$recvDepName))
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for ts, antBearing, fullID, recvDepName}
}
\description{
Plot tag ID vs time for all tags detected by site, coloured by antenna bearing
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
