% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSite.R
\name{plotSite}
\alias{plotSite}
\title{Plot all tags at a specified site}
\usage{
plotSite(data, sitename = unique(data$site))
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\description{
Plot tag ID vs time for all tags detected at a specified site, coloured by antenna
}
\examples{
plotSite(dat, sitename = "Piskwamish")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
