% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiteSig.R
\name{plotSiteSig}
\alias{plotSiteSig}
\title{Plot signal strength of all tags by a specified site}
\usage{
plotSiteSig(data, sitename)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags" or "alltagswithambigs", or a data.frame of detection data 
including at a minimum the variables antBearing, ts, lat, sig, fullID, recvDepName}

\item{sitename}{name of recvDepName}
}
\description{
Plot signal strength vs time for all tags detected at a specified site, coloured by antenna
}
\examples{
You can use either the tbl or the flat format for the siteTrans function, instructions to convert
a .motus file to both formats is below.
To access any tbl from .motus data saved on your computer:
file.name <- "data/project-sample.motus" ## replace with the full location of the sample dataset or your own project-XX.motus file
tmp <- dplyr::src_sqlite(file.name)
alltags <- tbl(motusSqlFile, "alltags")

To convert tbl to flat format:
alltags <- alltags \%>\% collect \%>\% as.data.frame

plot all tags for site Piskwamish
plotSiteSig(alltags, sitename = "Piskwamish")

Plot select tags for site Piskwamish 
plotSiteSig(filter(alltags, motusTagID \%in\% c(16037, 16039, 16035)), sitename = "Netitishi")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
