% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiteSig.R
\name{plotSiteSig}
\alias{plotSiteSig}
\title{Plot signal strength of all tags at a specified site}
\usage{
plotSiteSig(data, sitename = unique(data$site))
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\description{
Plot signal strength vs time for all tags detected at a specified site, coloured by antenna
}
\examples{
plotSiteSig(dat, sitename = "Piskwamish")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
