% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagSig.R
\name{plotTagSig}
\alias{plotTagSig}
\title{Plot signal strength of all detections for a specified tag by site}
\usage{
plotTagSig(data, motusTagID)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data
including at a minimum variables for motusTagID, sig, ts, antBearing, recvDeployLat, fullID, recvDeployName}

\item{motusTagID}{a numeric motusTagId to display in plot}
}
\description{
Plot signal strength vs time for specified tag, faceted by site (ordered by latitude) and coloured by antenna
}
\examples{
You can use either a selected tbl from .motus eg. "alltags", or a data.frame, instructions to convert a .motus file to all formats are below.
sql.motus <- tagme(176, new = TRUE, update = TRUE) # download and access data from project 176 in sql format
tbl.alltags <- tbl(sql.motus, "alltags") # convert sql file "sql.motus" to a tbl called "tbl.alltags"
df.alltags <- tbl.alltags \%>\% collect \%>\% as.data.frame() ## convert the tbl "tbl.alltags" to a data.frame called "df.alltags"

Plot signal strength of a specified tag using dataframe df.alltags
plotTagSig(df.alltags, motusTagID = 16047)

Plot signal strength of a specified tag using tbl file tbl.alltags
plotTagSig(tbl.alltags, motusTagID = 16035)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
