% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagSig.R
\name{plotTagSig}
\alias{plotTagSig}
\title{Plot signal strength of all detections for a specified tag by site}
\usage{
plotTagSig(data, motusTagID)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame of detection data 
including at a minimum variables for motusTagID, sig, ts, antBearing, recvDeployLat, fullID, recvDepName}

\item{motusTagID}{a numeric motusTagId to display in plot}
}
\description{
Plot signal strength vs time for specified tag, faceted by site (ordered by latitude) and coloured by antenna
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
