% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTagSig.R
\name{plotTagSig}
\alias{plotTagSig}
\title{Plot signal strength of all detections for a specified tag by site}
\usage{
plotTagSig(data, tag)
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\description{
Plot signal strength vs time for specified tag, faceted by site (ordered by latitude) and coloured by antenna
}
\examples{
plotTagSig(dat, tag = 171)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
