% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSiteDet.R
\name{simSiteDet}
\alias{simSiteDet}
\title{Create a dataframe of simultaneous detections at multiple sites}
\usage{
simSiteDet(data)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags" or "alltagswithambigs", or a data.frame of detection data 
including at a minimum the variables motusTagID, site, ts}
}
\description{
Creates a dataframe consisting of detections of tags that are detected at one or more receiver
at the same time.
}
\examples{
You can use either the tbl or the flat format for the siteTrans function, instructions to convert
a .motus file to both formats is below.
To access any tbl from .motus data saved on your computer:
file.name <- "data/project-sample.motus" ## replace with the full location of the sample dataset or your own project-XX.motus file
tmp <- dplyr::src_sqlite(file.name)
alltags <- tbl(motusSqlFile, "alltags")

To convert tbl to flat format:
alltags <- alltags \%>\% collect \%>\% as.data.frame

To get a data.frame of just simultaneous detections
tmp <- simSiteDet(alltags)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
