% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSiteDet.R
\name{simSiteDet}
\alias{simSiteDet}
\title{Create a dataframe of simultaneous detections at multiple sites}
\usage{
simSiteDet(data)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags", or a data.frame
of detection data including at a minimum variables for \code{motusTagID},
\code{recvDeployName}, \code{ts}}
}
\description{
Creates a dataframe consisting of detections of tags that are detected at two
or more receiver at the same time.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# convert sql file "sql_motus" to a tbl called "tbl_alltags"
library(dplyr)
tbl_alltags <- tbl(sql_motus, "alltags") 

# convert the tbl "tbl_alltags" to a data.frame called "df_alltags"
df_alltags <- tbl_alltags \%>\% 
  collect() \%>\% 
  as.data.frame()

# To get a data.frame of just simultaneous detections from a tbl file
# tbl_alltags
simSites <- simSiteDet(tbl_alltags)

# To get a data.frame of just simultaneous detections from a dataframe
# df_alltags
simSites <- simSiteDet(df_alltags)
}
