% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSiteDet.R
\name{simSiteDet}
\alias{simSiteDet}
\title{Create a dataframe of simultaneous detections at multiple sites}
\usage{
simSiteDet(data)
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\description{
Creates a dataframe consisting of detections of tags that are detected at one or more receiver
at the same time.
}
\examples{
simSiteDet(dat)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
