% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteSum.R
\name{siteSum}
\alias{siteSum}
\title{Summarize and plot detections of all tags by site}
\usage{
siteSum(data, site, fullID, ts, units = "hours")
}
\arguments{
\item{data}{dataframe of Motus detection data}

\item{units}{units to display time difference, defaults to "hours", options include "secs", "mins", "hours", "days", "weeks"}
}
\value{
a data.frame with these columns:
\itemize{
\item site: site
\item first_ts: time of first detection at specified site
\item last_ts: time of last detection at specified site
\item tot_ts: total amount of time between first and last detection at specified site, output in specified unit (defaults to "hours")
\item num.tags: total number of unique tags detected at specified site
\item num.det: total number of tag detections at specified site
}
}
\description{
Creates a summary of the first and last detection at a site, the length of time between first and last detection,
the number of tags, and the total number of detections at a site.  Plots total number of detections across all tags,
and total number of tags detected at each site.
}
\examples{
site_summary <- siteSum(dat, units = "mins")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
