% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteSumDaily.R
\name{siteSumDaily}
\alias{siteSumDaily}
\title{Summarize daily detections of all tags by site}
\usage{
siteSumDaily(data, units = "hours")
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltagsGPS", or a
data.frame of detection data including at a minimum variables for
\code{motusTagID}, \code{sig}, \code{recvDeployName}, \code{ts}}

\item{units}{units to display time difference, defaults to "hours", options
include "secs", "mins", "hours", "days", "weeks"}
}
\value{
a data.frame with these columns:
\itemize{
\item recvDeployName: site name of deployment
\item date: date that is being summarized
\item first_ts: time of first detection on specified "date" at "recvDeployName"
\item last_ts: time of last detection on specified "date" at "recvDeployName"
\item tot_ts: total amount of time between first and last detection at
"recvDeployName" on "date, output in specified unit (defaults to "hours")
\item num.tags: total number of unique tags detected at "recvDeployName", on "date"
\item num.det: total number of detections at "recvDeployName", on "date"
}
}
\description{
Creates a summary of the first and last daily detection at a site, the length
of time between first and last detection, the number of tags, and the total
number of detections at a site for each day. Same as \code{siteSum()}, but daily
by site.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE, update = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# convert sql file "sql_motus" to a tbl called "tbl_alltags"
library(dplyr)
tbl_alltags <- tbl(sql_motus, "alltagsGPS")

# convert the tbl "tbl_alltags" to a data.frame called "df_alltags"
df_alltags <- tbl_alltags \%>\% 
  collect() \%>\% 
  as.data.frame() 

# Create site summaries for all sites within detection data with time in
# minutes using tbl file tbl_alltags
daily_site_summary <- siteSumDaily(tbl_alltags, units = "mins")

# Create site summaries for only select sites with time in minutes using tbl
# file tbl_alltags
sub <- filter(tbl_alltags, recvDeployName \%in\% c("Niapiskau", "Netitishi", 
                                                 "Old Cut", "Washkaugou"))
daily_site_summary <- siteSumDaily(sub, units = "mins")

# Create site summaries for only a select species, Red Knot, with default
# time in hours using data frame df_alltags
daily_site_summary <- siteSumDaily(filter(df_alltags,
                                          speciesEN == "Red Knot"))
}
