% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteTrans.R
\name{siteTrans}
\alias{siteTrans}
\title{Summarize transitions between sites for each tag}
\usage{
siteTrans(data)
}
\arguments{
\item{data}{dataframe of Motus detection data containing at a minimum fullID, ts, lat, lon}
}
\value{
a data.frame with these columns:
\itemize{
\item fullID: fullID of Motus registered tag
\item ts.x: time of last detection of tag at site.x ("departure" time)
\item lat.x: latitude of site.x
\item lon.x: longitude of site.x
\item site.x: first site in transition pair (the "departure" site)
\item ts.y: time of first detection of tag at site.y ("arrival" time)
\item lat.y: latitude of site.y
\item lon.y: longitude of site.y
\item site.y: second site in transition pair (the "departure" site)
\item tot_ts: length of time between ts.x and ts.y (in seconds)
\item dist: total straight line distance between site.x and site.y (in metres), see latLonDist function in sensorgnome package for details
\item rate: overall rate of movement (tot_ts/dist), in metres/second
\item bearing: bearing between first and last detection sites, see bearing function in geosphere package for more details
\item rhumbline_bearing: rhumbline bearing between first and last detection sites, see bearingRhumb function in geosphere package for more detail
}
}
\description{
Creates a dataframe of transitions between sites; detections are ordered by detection time, then "transitions"
are identified as the period between the final detection at site x (possible "departure"), and the first detection
(possible "arrival") at site y (ordered chronologically). Each row contains the last detection time and lat/lon
of site x, first deteciton time and lat/lon of site y, distance between the site pair, time between detections,
rate of movement between detections, bearing and rhumbline bearing between site pair.
}
\examples{
transitions <- siteTrans(dat)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
