% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvRunsForReceiverProject.R
\name{srvRunsForReceiverProject}
\alias{srvRunsForReceiverProject}
\title{get the runs for a receiver project from the data server}
\usage{
srvRunsForReceiverProject(projectID, batchID, runID = 0, countOnly = FALSE)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{batchID}{integer scalar motus batch ID}

\item{runID}{integer scalar ID of latest run already obtained.
Default: 0, meaning none.}

\item{countOnly}{logical; if TRUE, return only the cound of available batches.
Default: FALSE.}
}
\value{
data.frame with these columns:
\itemize{
\item runID
\item batchIDbegin
\item batchIDend
\item motusTagID
\item ant
\item len
}
}
\description{
get the runs for a receiver project from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}

