% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srvRunsForTagProject.R
\name{srvRunsForTagProject}
\alias{srvRunsForTagProject}
\title{get the runs for a tag project in a batch from the data server}
\usage{
srvRunsForTagProject(projectID, batchID, runID = 0)
}
\arguments{
\item{projectID}{integer scalar motus project ID}

\item{batchID}{integer scalar motus batch ID}

\item{runID}{integer scalar ID of latest run already obtained.
Default: 0, meaning none.}
}
\value{
a data.frame with these columns:
\itemize{
   \item runID
   \item batchIDbegin
   \item tsBegin
   \item tsEnd
   \item done
   \item motusTagID
   \item ant
   \item len
}
}
\description{
get the runs for a tag project in a batch from the data server
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
