% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunRiseSet.R
\name{SunRiseSet}
\alias{SunRiseSet}
\title{Obtain sunrise and sunset times}
\usage{
SunRiseSet(data)
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags" or "alltagswithambigs", or a data.frame of detection data 
including at a minimum the variables ts, lat, lon}
}
\value{
the original dataframe provided, with the following additional columns:
- sunrise: sunrise time for the date and location provided by ts and lat/lon per row
- sunset: sunset time for the date and location provided by ts and lat/lon per row
}
\description{
Creates and adds a sunrise and sunset column to a data.frame containing latitude, longitude, and a POSIXct date/time
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
