% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunRiseSet.R
\name{sunRiseSet}
\alias{sunRiseSet}
\title{Obtain sunrise and sunset times}
\usage{
sunRiseSet(data, lat = "recvDeployLat", lon = "recvDeployLon", ts = "ts")
}
\arguments{
\item{data}{a selected table from .motus detection data, eg. "alltags", or a data.frame of detection data
including at a minimum variables for date/time, latitude, and longitude}

\item{lat}{variable with latitude values, defaults to recvDeployLat}

\item{lon}{variable with longitude values, defaults to recvDeployLon}

\item{ts}{variable with time in UTC as numeric or POSIXct, defaults to ts}
}
\value{
the original dataframe provided, with the following additional columns:
\itemize{
\item sunrise: sunrise time for the date and location provided by ts and lat/lon per row
\item sunset: sunset time for the date and location provided by ts and lat/lon per row
}
}
\description{
Creates and adds a sunrise and sunset column to a data.frame containing latitude, longitude,
and a date/time as POSIXct or numeric.
}
\examples{
# You can use either a selected tbl from .motus eg. "alltags", or a
# data.frame, instructions to convert a .motus file to all formats are below.

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# OR use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))

# convert sql file "sql.motus" to a tbl called "tbl.alltags"
library(dplyr)
tbl.alltags <- tbl(sql.motus, "alltags") 

# convert the tbl "tbl.alltags" to a data.frame called "df.alltags"
df.alltags <- tbl.alltags \%>\% 
  collect() \%>\% 
  as.data.frame() 

# Add sunrise and sunset columns to a data.frame from alltags
sun <- sunRiseSet(df.alltags)

# Get sunrise and sunset information from tbl.alltags using gps lat/lon
# Note this only works if there are non-NA values in gpsLat/gpsLon
\dontrun{sun <- sunRiseSet(tbl.alltags, lat = "gpsLat", lon = "gpsLon")}
}
