% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunRiseSet.R
\name{SunRiseSet}
\alias{SunRiseSet}
\title{Obtain sunrise and sunset times}
\usage{
SunRiseSet(data)
}
\arguments{
\item{data}{dataframe of Motus detection data, or dataframe containing latitude, longitude, and POSIXct date/time}
}
\value{
the original dataframe provided, with the following additional columns:
- sunrise: sunrise time for the date and location provided by ts and lat/lon per row
- sunset: sunset time for the date and location provided by ts and lat/lon per row
}
\description{
Creates and adds a sunrise and sunset column to a dataframe containing latitude, longitude, and a POSIXct date/time
}
\examples{
dat <- SunRiseSet(dat) ## adds sunrise and sunset columns to dataframe "dat"
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
