% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagme.R
\name{tagme}
\alias{tagme}
\title{update a motus tag detection database}
\usage{
tagme(projRecv, update = TRUE, new = FALSE, dir = getwd(),
  countOnly = FALSE, forceMeta = FALSE)
}
\arguments{
\item{projRecv}{integer scalar project code from motus.org, *or*
character scalar receiver serial number}

\item{update}{boolean scalar: should any new data be downloaded and merged?
default: TRUE, unless this is a new database (in which case you must
specify \code{update=TRUE} explicitly).}

\item{new}{logical scalar: is this a new database?  Default: FALSE
You have to specify \code{new=TRUE} if you want a new local copy of the
database to be created.  Otherwise, \code{tagme()} assumes the database
already exists, and will stop with an error if it cannot find it
in the current directory.  This is mainly to prevent inadvertent
downloads of large amounts of data that you already have!}

\item{countOnly}{logical scalar: if FALSE, the default, then do
requested database updates.  Otherwise, return a count of items
that would need to be transferred in order to update the
database.}

\item{forceMeta}{logical scalar: if true, re-get metadata for tags and
receivers, even if we already have them.}

\item{dir:}{path to the folder where you are storing databases
Default: the current directory; i.e. \code{getwd()}}
}
\value{
a dplyr::src_sqlite for the (possibly updated) database, or a list
of counts if \code{countOnly==TRUE}
}
\description{
update a motus tag detection database
}
\examples{

## create and open a local tag database for motus project 14 in the
## current directory

# t = tagme(14, new=TRUE)

## update and open the local tag database for motus project 14;
## it must already exist and be in the current directory

# t = tagme(14, update=TRUE)

## update and open the local tag database for a receiver;
## it must already exist and be in the current directory

# t = tagme("SG-1234BBBK4567", update=TRUE)

## open the local tag database for a receiver, without
## updating it

# t = tagme("SG-1234BBBK4567")

## open the local tag database for a receiver, but
## tell 'tagme' that it is in a specific directory

# t = tagme("SG-1234BBBK4567", dir="Projects/gulls")

## update all existing project and receiver databases in \\code{dir}
# tagme()

}
\seealso{
\code{\link{tellme}}, which is a synonym for \code{tagme(..., update=TRUE, countOnly=TRUE)}
}
\author{
John Brzustowski \email{jbrzusto@REMOVE_THIS_PART_fastmail.fm}
}
