% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagSum.R
\name{tagSum}
\alias{tagSum}
\title{General summary of detections for each tag}
\usage{
tagSum(data)
}
\arguments{
\item{data}{dataframe of Motus detection data}
}
\value{
a data.frame with these columns:
\itemize{
\item fullID: fullID of Motus registered tag
\item first_ts: time of first detection of tag
\item last_ts: time of last detection of tag
\item first_site: first detection site of tag
\item last_site: last deteciton site of tag
\item lat.x: latitude of first deteciton site of tag
\item lon.x: longitude of first deteciton site of tag
\item lat.y: latitude of last deteciton site of tag
\item lon.y: longitude of last deteciton site of tag
\item tot_ts: length of time between first and last detection of tag (in seconds)
\item dist: total straight line distance between first and last detection site (in metres), see latLonDist function in sensorgnome package for details
\item rate: overall rate of movement (tot_ts/dist), in metres/second
\item bearing: bearing between first and last detection sites, see bearing function in geosphere package for more details
\item rhumbline_bearing: rhumbline bearing between first and last detection sites, see bearingRhumb function in geosphere package for more detail
}
}
\description{
Creates a summary for each tag of it's first and last detection time, first and last detection site,
length of time between first and last detection,  straight line distance between first and last detection site,
rate of movement, bearing, and rhumbline bearing
}
\examples{
tag_summary <- tagSum(dat)
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
