% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagSumSite.R
\name{tagSumSite}
\alias{tagSumSite}
\title{Summarize detections of all tags by site}
\usage{
tagSumSite(data, units = "hours")
}
\arguments{
\item{data}{dataframe of Motus detection data}

\item{units}{units to display time difference, defaults to "hours", options include "secs", "mins", "hours", "days", "weeks"}
}
\description{
Creates a summary for each tag of it's first and last detection time at each site,
length of time between first and last detection of each site, and total number of detections at each site
}
\examples{
tag_site_summary <- tagSumSite(dat, units = "mins")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
