% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagSumSite.R
\name{tagSumSite}
\alias{tagSumSite}
\title{Summarize detections of all tags by site}
\usage{
tagSumSite(data, units = "hours")
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltagsGPS", or a
data.frame of detection data including at a minimum variables for
motusTagID, fullID, recvDeployName, ts, recvDeployLat, recvDeployLon,
gpsLat, gpsLon}

\item{units}{units to display time difference, defaults to "hours", options
include "secs", "mins", "hours", "days", "weeks"}
}
\description{
Creates a summary for each tag of it's first and last detection time at each
site, length of time between first and last detection of each site, and total
number of detections at each site.
}
\examples{
# You can use either a selected tbl from .motus eg. "alltagsGPS", or a
# data.frame, instructions to convert a .motus file to all formats are below.

# download and access data from project 176 in sql format
# usename and password are both "motus.sample"
\dontrun{sql.motus <- tagme(176, new = TRUE, update = TRUE)}

# use example sql file included in `motus`
sql.motus <- tagme(176, update = FALSE, 
                   dir = system.file("extdata", package = "motus"))

# convert sql file "sql.motus" to a tbl called "tbl.alltags"
library(dplyr)
tbl.alltags <- tbl(sql.motus, "alltagsGPS") 

# convert the tbl "tbl.alltags" to a data.frame called "df.alltags"
df.alltags <- tbl.alltags  \%>\% 
  collect() \%>\% 
  as.data.frame() 

# Create tag summaries for all tags within detection data with time in
# minutes with tbl file tbl.alltags
tag_site_summary <- tagSumSite(tbl.alltags, units = "mins")

# Create tag summaries for only select tags with time in default hours with
# data.frame df.alltags
tag_site_summary <- tagSumSite(filter(df.alltags, 
                                      motusTagID \%in\% c(16047, 16037, 16039)))

# Create tag summaries for only a select species with data.frame df.alltags
tag_site_summary <- tagSumSite(filter(df.alltags, speciesEN == "Red Knot"))
}
