% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToSunriset.R
\name{timeToSunriset}
\alias{timeToSunriset}
\title{Obtain time to and from sunrise/sunset for a dataframe containing POSIXct times}
\usage{
timeToSunriset(data, units = "hours")
}
\arguments{
\item{data}{a selected table from .motus data, eg. "alltags" or "alltagswithambigs", or a data.frame of detection data 
including at a minimum the variables ts, lat, lon}

\item{units}{units to display time difference, defaults to "hours", options include "secs", "mins", "hours", "days", "weeks"}
}
\value{
the original dataframe provided, with the following additional columns:
\itemize{
\item sunrise: sunrise time for the date and location provided by ts and lat/lon per row
\item sunset: sunset time for the date and location provided by ts and lat/lon per row
\item ts_to_set: time to next sunset after "ts", units default to "hours"
\item ts_since_set: time to previous sunset since "ts", units default to "hours"
\item ts_to_rise: time to next sunrise after "ts", units default to "hours"
\item ts_since_rise: time to previous sunrise since "ts", units default to "hours"
}
}
\description{
Creates and adds columns for time to, and time from sunrise/sunset based on a column of POSIXct dates/times
dataframe must contain latitude, longitude, and a POSIXct date/time.
}
\examples{
You can use either the tbl or the flat format for the siteTrans function, instructions to convert
a .motus file to both formats is below.
To access any tbl from .motus data saved on your computer:
file.name <- "data/project-sample.motus" ## replace with the full location of the sample dataset or your own project-XX.motus file
tmp <- dplyr::src_sqlite(file.name)
alltags <- tbl(motusSqlFile, "alltags")

To convert tbl to flat format:
alltags <- alltags \%>\% collect \%>\% as.data.frame

get sunrise and sunset information with units in minutes
sunrise <- timeToSunriset(alltags, units = "mins")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
