% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToSunriset.R
\name{timeToSunriset}
\alias{timeToSunriset}
\title{Obtain time to and from sunrise/sunset}
\usage{
timeToSunriset(
  df_src,
  lat = "recvDeployLat",
  lon = "recvDeployLon",
  ts = "ts",
  units = "hours",
  data
)
}
\arguments{
\item{df_src}{Data frame, SQLite connection, or SQLite table. An SQLite
connection would be the result of \code{tagme(XXX)} or
\code{DBI::dbConnect(RSQLite::SQLite(), "XXX.motus")}; an SQLite table would be
the result of \code{dplyr::tbl(tags, "alltags")}; a data frame could be the
result of \code{dplyr::tbl(tags, "alltags") \%>\% dplyr::collect()}.}

\item{lat}{Character. Name of column with latitude values, defaults to
\code{recvDeployLat}}

\item{lon}{Character. Name of column with longitude values, defaults to
\code{recvDeployLon}}

\item{ts}{Character. Name of column with time as numeric or POSIXct, defaults
to \code{ts}}

\item{units}{Character. Units to display time difference, defaults to
"hours", options include "secs", "mins", "hours", "days", "weeks".}

\item{data}{Defunct, use \code{src}, \code{df_src}, or \code{df} instead.}
}
\value{
Original data (as a flat data frame), with the following additional
columns:
\itemize{
\item \code{sunrise} - Time of sunrise in \strong{UTC} for that row's date and location
\item \code{sunset} - Time of sunset in \strong{UTC} for that row's date and location
\item \code{ts_to_set} - Time to next sunset, in \code{units}
\item \code{ts_since_set} - Time to previous sunset, in \code{units}
\item \code{ts_to_rise} - Time to next sunrise after, in \code{units}
\item \code{ts_since_rise} - Time to previous sunrise, in \code{units}
}
}
\description{
Create and add columns for time to and time since sunrise/sunset to tag data.
Can take a motus database table, but will always return a collected data
frame. Requires data containing at least latitude, longitude, and time.
}
\details{
Uses \code{sunRiseSet()} to perform sunrise/sunset calculates, see \code{?sunRiseSet}
for details regarding how local dates are assessed from UTC timestamps.
}
\examples{
# Download sample project 176 to .motus database (username/password are "motus.sample")
\dontrun{sql_motus <- tagme(176, new = TRUE)}

# Or use example data base in memory
sql_motus <- tagmeSample()

# Get sunrise and sunset information for alltags view with units in minutes
sunrise <- timeToSunriset(sql_motus, units = "mins")
}
