% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToSunriset.R
\name{timeToSunriset}
\alias{timeToSunriset}
\title{Obtain time to and from sunrise/sunset for a dataframe containing POSIXct times}
\usage{
timeToSunriset(data, units = "hours")
}
\arguments{
\item{data}{dataframe of Motus detection data, or dataframe containing latitude, longitude, and POSIXct date/time}

\item{units}{units to display time difference, defaults to "hours", options include "secs", "mins", "hours", "days", "weeks"}
}
\value{
the original dataframe provided, with the following additional columns:
\itemize{
\item sunrise: sunrise time for the date and location provided by ts and lat/lon per row
\item sunset: sunset time for the date and location provided by ts and lat/lon per row
\item ts_to_set: time to next sunset after "ts", units default to "hours"
\item ts_since_set: time to previous sunset since "ts", units default to "hours"
\item ts_to_rise: time to next sunrise after "ts", units default to "hours"
\item ts_since_rise: time to previous sunrise since "ts", units default to "hours"
}
}
\description{
Creates and adds columns for time to, and time from sunrise/sunset based on a column of POSIXct dates/times
dataframe must contain latitude, longitude, and a POSIXct date/time.
}
\examples{
dat <- timeToSunriset(dat, units = "mins")
}
\author{
Zoe Crysler \email{zcrysler@gmail.com}
}
