% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMotusDb.r
\name{updateMotusDb}
\alias{updateMotusDb}
\title{Update motus sqlite file}
\usage{
updateMotusDb(rv, src)
}
\arguments{
\item{rv}{return value}

\item{src}{sqlite database source}
}
\value{
rv
}
\description{
Ensures that the motus sqlite file is up-to-date to support the current
version of the package. Relies on the \code{sql_versions} internal data frame
to run the SQL on the basis of date.
}
\details{
This function adds a new admInfo table in the motus sqlite file that keeps
track of the date at which the last correction was applied. The updateMotusDb
function only exectute sql commands added since the last correction.

To insert a new version update modify and run the ./data-raw/updatesql.R
script to add new sql commands to the internal data frame.

\itemize{
\item date: date at which the sql update record was added (default current
timestamp)
\item sql: sql string to execute (you should minimize the risk of database
errors by using IF EXISTS or DROP as appropriate prior to your command)
\item descr: description of the update, printed for the user during the
update process
}

This function is called at the end of the \code{\link{ensureDBTables}}
function. i.e., it will be called each time that a motus file is opened.
}
\author{
Denis Lepage \email{dlepage@bsc-eoc.org}
}
\keyword{internal}
