% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRunsFilter.R
\name{writeRunsFilter}
\alias{writeRunsFilter}
\title{Write to the local database the probabilities associated with runs for a filter}
\usage{
writeRunsFilter(
  src,
  filterName,
  motusProjID = NA,
  df,
  overwrite = TRUE,
  delete = FALSE
)
}
\arguments{
\item{src}{dplyr sqlite src, as returned by \code{dplyr::src_sqlite()}}

\item{filterName}{unique name given to the filter}

\item{motusProjID}{optional project ID attached to the filter in order to share with other users of the same project.}

\item{df}{dataframe containing the runID, the motusTagID and probability values to save in the local database}

\item{overwrite}{boolean. When TRUE ensures that existing records matching the same filterName
and runID get replaced}

\item{delete}{boolean. When TRUE, removes all existing filter records associated with the filterName
and re-inserts the ones contained in the dataframe. This option should be used if the dataframe
provided contains the entire set of filters you want to save.}
}
\value{
a dplyr sqlite object refering to the filter created
}
\description{
Write to the local database the probabilities associated with runs for a filter
}
